% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{build_dif_tables}
\alias{build_dif_tables}
\title{Combine individual DIF results}
\usage{
build_dif_tables(dif_summaries, save_at)
}
\arguments{
\item{dif_summaries}{named list of summary_dif() return objects; the list
elements must be named after their DIF variable}

\item{save_at}{character string; indicates the folder location where the
summaries are stored on the hard drive. Please note that the
path is relative to the current working path set by here::i_am()}
}
\description{
Combine the individual DIF analysis results to excel tables that can be
used for the technical reports
}
