% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{build_dif_tr_tables}
\alias{build_dif_tr_tables}
\title{Combine individual DIF results}
\usage{
build_dif_tr_tables(
  dif_summaries,
  save = TRUE,
  path_results = here::here("Results"),
  path_table = here::here("Tables"),
  overwrite = FALSE
)
}
\arguments{
\item{dif_summaries}{named list of dif_summary() return objects; the list
elements must be named after their DIF variable}

\item{save}{logical; whether results shall be saved to hard drive}

\item{path_results}{string; indicates the folder location where the DIF
estimates are stored on the hard drive; please note that the path is relative
to the current working path set by here::i_am()}

\item{path_table}{string; indicates the folder location where the summaries
are stored on the hard drive; please note that the path is relative to the
current working path set by here::i_am()}

\item{overwrite}{logical; whether to overwrite existing file when saving table}
}
\description{
Combine the individual DIF analysis results to excel tables that can be
used for the technical reports
}
