% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{calculate_link_parameters}
\alias{calculate_link_parameters}
\title{calculate link constant and link error for anchor group or items design}
\usage{
calculate_link_parameters(
  vars_curr,
  vars_prev,
  vars_link = NULL,
  select_curr,
  select_prev,
  select_link = NULL,
  scoring_curr = NULL,
  scoring_prev = NULL,
  scoring_link = NULL,
  xsi_curr,
  xsi_prev,
  xsi_link = NULL,
  anchors = NULL,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{vars_curr}{data.frame; contains information about items of current
measurement wave with items as rows; includes variable 'item' containing item
names and may include variable with item scoring}

\item{vars_prev}{data.frame; contains information about items of previous
measurement wave with items as rows; includes variable 'item' containing item
names and may include variable with item scoring}

\item{vars_link}{data.frame; contains information about items of link study
with items as rows; includes variable 'item' containing item names and may
include variable with item scoring; required if anchor group design is used}

\item{select_curr}{character; contains name of logical variable in vars_curr
identifying the item set for the current measurement wave}

\item{select_prev}{character; contains name of logical variable in vars_prev
identifying the item set for the previous measurement wave}

\item{select_link}{character; contains name of logical variable in vars_link
identifying the item set for the link study; required if anchor group design
is used}

\item{scoring_curr}{string; defines name of numerical variable in vars_curr
that contains the scoring factor to be applied to the loading matrix for the
current measurement point; can be NULL for the Rasch model}

\item{scoring_prev}{string; defines name of numerical variable in vars_prev
that contains the scoring factor to be applied to the loading matrix for the
previous measurement point; can be NULL for the Rasch model}

\item{scoring_link}{string; defines name of numerical variable in vars_link
that contains the scoring factor to be applied to the loading matrix for the
link study; can be NULL for the Rasch model; optional variable for anchor
group design}

\item{xsi_curr}{matrix; item difficulty parameters of current measurement}

\item{xsi_prev}{matrix; item difficulty parameters of previous measurement}

\item{xsi_link}{matrix; item difficulty parameters of link sample}

\item{anchors}{character; data.frame with two columns including the link
items; for anchor item designs the first column refers to the previous
measurement time point and the second column to the current measurement
time point; for anchor group designs the first column refers to the main
sample (for both time points) and the second column refers to the link
sample; if NULL, all common items are used}

\item{warn}{logical; whether to print warnings}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\value{
list(const, const.err, anchors_prev, anchors_curr, xsi_curr.linked)
}
\description{
calculate link constant and link error for anchor group or items design
}
