% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{calculate_num_cat}
\alias{calculate_num_cat}
\title{Calculate number of categories for each item}
\usage{
calculate_num_cat(vars, poly_items, select_dich = "dich")
}
\arguments{
\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{poly_items}{list; contains character vector with subitems for each
polytomous item, name of the vector is the name of the polytomous item (e.g.
poly_items = list(poly1 = c("subitem1", "subitem2"), poly2 = c("subitem1", "subitem2")))}

\item{select_dich}{string; name of logical variable in vars that identifies
all dichotomous items for which number of categories is set to 1}
}
\value{
numeric vector with number of subitems/categories for all items
}
\description{
Calculate number of categories for each item
}
