% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{check_dif_anchor}
\alias{check_dif_anchor}
\title{Test measurement invariance / DIF over time}
\usage{
check_dif_anchor(
  resp_previous,
  resp_current,
  resp_link_sample = NULL,
  anchors = NULL,
  mvs = NULL,
  return = TRUE,
  valid,
  scoring = "scoring",
  vars,
  items = NULL,
  select_previous,
  select_current
)
}
\arguments{
\item{resp_previous}{data.frame with responses of first measurement wave;
contains ID_t; anchor items are named to fit the current analysis}

\item{resp_current}{data.frame with responses of second measurement wave;
contains ID_t}

\item{resp_link_sample}{data.frame with responses of link sample;
contains ID_t; required if anchor group design is used}

\item{anchors}{data.frame with two columns including the link items
in the two waves; if NULL, all common items are used}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{items}{character; contains name of logical variable in vars
identifying the item set for the both time points}

\item{select_previous}{character; contains name of logical variable in vars
identifying the item set for the previous measurement time point; needed to
build tables}

\item{select_current}{character; contains name of logical variable in vars
identifying the item set for the current measurement time point; needed to
build tables}
}
\value{
list with DIF model(s), DIF summary and anchor items
}
\description{
Test measurement invariance / DIF over time
}
