% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_folder}
\alias{check_folder}
\title{Check if folder exists and if not, create new one}
\usage{
check_folder(path)
}
\arguments{
\item{path}{string; indicates the folder location that shall be checked;
please note that the path is relative to the current working path set by
here::i_am()}
}
\description{
Check if folder exists and if not, create new one
}
