% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_items}
\alias{check_items}
\title{Check item names for duplicates}
\usage{
check_items(items)
}
\arguments{
\item{items}{character vector with item names}
}
\description{
Check item names for duplicates
}
