% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{check_link_dimensionality}
\alias{check_link_dimensionality}
\title{Test unidimensionality over time}
\usage{
check_link_dimensionality(
  resp_previous,
  resp_current,
  resp_link_sample,
  vars,
  select_common,
  scoring,
  dim = "link",
  mvs,
  irtmodel = "PCM2",
  maxiter,
  snodes,
  verbose = FALSE,
  valid
)
}
\arguments{
\item{resp_previous}{data.frame with responses of first measurement wave,
a person identifier and a variable indicating valid cases; may also
include a WLE as given in wid}

\item{resp_current}{data.frame with responses of second measurement wave,
a person identifier and a variable indicating valid cases}

\item{resp_link_sample}{data.frame with responses of link sample,
a person identifier and a variable indicating valid cases; variables of
the link sample have to be adapted to the item names and the item scoring
in the previous and current measurement}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments: contains an
item identifier for all items in the link sample and a dimension identifier
for the time points; please note that, in the anchor items design, the
common items between the time points must be named the same and, therefore,
not be doubled in vars}

\item{select_common}{character; indicates the variable in vars
selecting the items that were administered in the link sample (anchor
groups design) or at both time points (anchor items design -- note that
anchor items only appear once!)}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to the link constant (e.g., only
half the link constant is needed if a PC item is scored 0.5)}

\item{dim}{character; indicates the variable in vars defining the dimension
of the items; anchor groups design: the time points form distinct
dimensions; anchor items design: the common items across time points and
the time point unique items form unique dimensions, respectively}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{irtmodel}{string; "1PL" for Rasch, "2PL" for 2PL, "PCM2" for PCM and
"GPCM" for GPCM analysis}

\item{maxiter}{max iterations as passed to the TAM function}

\item{snodes}{snodes as passed to the TAM function}

\item{verbose}{verbose as passed to the TAM function}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}
}
\description{
Test unidimensionality over time
}
