% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_numerics}
\alias{check_numerics}
\title{Check numeric variable(s) for correctness}
\usage{
check_numerics(
  df,
  name_df = "<unknown>",
  numerics = NULL,
  check_invalid = FALSE,
  dich = FALSE
)
}
\arguments{
\item{df}{data.frame; contains at least the indicated numeric variable(s)}

\item{name_df}{string; contains name of df}

\item{numerics}{character vector; contains names of variables that shall be
checked for correctness; if NULL, all items will be used}

\item{check_invalid}{logical; whether to check df for invalid values}

\item{dich}{logical; whether to check items for non-dichotomous values}
}
\value{
NULL invisibly
}
\description{
Check numeric variable(s) for correctness
}
