% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_variables}
\alias{check_variables}
\title{Check variables in df for existence}
\usage{
check_variables(df, name_df = "<unknown>", variables)
}
\arguments{
\item{df}{data.frame}

\item{name_df}{string; contains name of data.frame}

\item{variables}{character vector; contains names of variables that shall be
checked for inclusion in df}
}
\value{
NULL invisibly
}
\description{
Check variables in df for existence
}
