% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{collapse_response_categories_with_rules}
\alias{collapse_response_categories_with_rules}
\title{Collapse response categories with existing collapsing rules}
\usage{
collapse_response_categories_with_rules(
  resp,
  vars,
  select = "poly",
  per_cat = 200,
  rules = NULL,
  rules_file = NULL
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in \{0, 1\} for binary data and y in \{0, 1, ... k-1\} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{character; indicates the logical variable in vars which
contains the item names of the polytomous items}

\item{per_cat}{integer; minimum number of persons per category; defaults to 200}

\item{rules}{data.frame; collapsing rules for items;
same format as the excel file saved via \code{save}; must include three
variables (original_item, scoring, collapsed_item)}

\item{rules_file}{character; path to excel file with collapsing rules created
via \code{save}}
}
\description{
Collapse response categories with existing collapsing rules
}
