% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_mv}
\alias{convert_mv}
\title{Convert user-defined missing values to NAs}
\usage{
convert_mv(resp, variables = NULL, mvs = NULL)
}
\arguments{
\item{resp}{data.frame containing the responses, scored items,
sociodemographic and design variables}

\item{variables}{character vector with names of all competence variables}

\item{mvs}{numeric vector including all user-defined missing values}
}
\value{
data.frame like resp, but without user-defined mvs
}
\description{
User defined missing values (usually coded as negative numbers) must be
converted to NAs so that they do not break the IRT analysis
}
