% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{create_table}
\alias{create_table}
\title{Create table for all possible response categories (even if n = 0)}
\usage{
create_table(response)
}
\arguments{
\item{response}{responses for one item}
}
\value{
table
}
\description{
Create table for all possible response categories (even if n = 0)
}
