% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat4}
\alias{dat4}
\title{Polytomous NEPS data with multi-stage design}
\format{
A list with two data frames including the simulated responses
(\code{resp}) and the characteristics of the variables (\code{vars}).

\code{resp} is a data frame with 3,710 rows and 145 columns:
\describe{
\item{ID_t}{Unique identifier for respondent.}
\item{sex}{Sex with 0 = male and 1 = female.}
\item{mig}{Migrant background with 0 = no and 1 = yes.}
\item{school}{School type with 0 = lower, 1 = intermediate, and 1 = upper.}
\item{books}{Binary book count with 0 = less than 100, 1 = more than 100.}
\item{age}{Age in years.}
\item{t34005a}{Raw school variable with categories 1 to 6.}
\item{t400500}{Raw migration variable with categories 1 to 10.}
\item{testy}{Test year.}
\item{testm}{Test month.}
\item{birthy}{Birth year.}
\item{birthm}{Birth month.}
\item{rotation}{Test position with 0 = first and 1 = second.}
\item{sca14XXX_c}{Scored item responses 0 = incorrect and 1 = correct.}
\item{sca14XXX}{Unscored item responses for each (sub)item with values of
1 to 4 representing the selected response option.}
\item{stage1}{A logical identifying respondents who were administered the first stage.}
\item{stage2}{A logical identifying respondents who were administered the second stage.}
\item{stage3}{A logical identifying respondents who were administered the third stage.}
\item{valid}{A logical vector identifying valid respondents.}
}

\code{vars} is a data frame with 128 rows and 17 columns:
\describe{
\item{item}{Name of the item.}
\item{raw}{A logical distinguishing scored (\code{FALSE}) and unscored (\code{TRUE})
items.}
\item{dich}{A logical identifying scored dichotomous (\code{TRUE}) and
other (\code{FALSE}) items.}
\item{poly}{A logical identifying scored polytomous (\code{TRUE}) and other
(\code{FALSE}) items.}
\item{mixed}{A logical identifying scored (\code{TRUE}) and other (\code{FALSE})
items.}
\item{type}{A string identifying the item type as simple multiple-choice
(\code{MC}) or complex multiple-choice (\code{CMC}).}
\item{scoring}{A number with the scoring weight in the item response
analyses.}
\item{num_cat}{A number with the number of response categories.}
\item{component}{Contains the two components KOS and KAS.}
\item{correct}{A number between 1 and 4 representing the correct response
option for each unscored (sub)item.}
\item{stage1}{A logical identifying the items administered in the first stage.}
\item{stage2}{A logical identifying the items administered in the second stage.}
\item{stage3}{A logical identifying the items administered in the third stage.}
\item{level1}{A logical identifying the items corresponding to level 1.}
\item{level2}{A logical identifying the items corresponding to level 2.}
\item{level3}{A logical identifying the items corresponding to level 3.}
\item{pos}{A number with the item position in the test.}
}
}
\usage{
dat4
}
\description{
The simulated data resembles the NEPS data structure for a multi-sateg test with 26
dichotomous and 8 polytomous items presented in three stages.
}
\keyword{datasets}
