% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desc_abs}
\alias{desc_abs}
\title{Show frequency (absolute) of answers of list of variables}
\usage{
desc_abs(dat, desc, valid = NULL)
}
\arguments{
\item{dat}{data.frame; contains sociodemographic and booklet variables}

\item{desc}{character vector; contains the name(s) of the variable(s) that
shall be used in analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}
}
\value{
list with  frequency (absolute) of answers for each variable in desc.
}
\description{
Show frequency (absolute) of answers of list of variables
}
