% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desc_attr}
\alias{desc_attr}
\title{Show attributes of list of variables}
\usage{
desc_attr(dat, desc, min.val = 3)
}
\arguments{
\item{dat}{data.frame with item responses (user-defined missing values)
and logical grouping variables, named after the grouping
factor (e.g. "easy" and "difficult" in the case of testlet.}

\item{desc}{character. contains names of variables for descriptive analysis.}

\item{min.val}{minimum number of valid values;
if negative, set to the default of 3}
}
\value{
list with attributes for each variable in desc.
}
\description{
Show attributes of list of variables
}
