% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desc_nom}
\alias{desc_nom}
\title{Descriptives of nominal variables}
\usage{
desc_nom(
  dat,
  desc,
  valid = NULL,
  digits = 1,
  show_absolute = TRUE,
  show_percentage = TRUE
)
}
\arguments{
\item{dat}{data.frame; contains sociodemographic and booklet variables}

\item{desc}{character vector; contains the name(s) of the variable(s) that
shall be used in analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{digits}{integer; number of decimals for rounding of percentages}

\item{show_absolute}{logical; whether to show frequency in absolute numbers}

\item{show_percentage}{logical; whether to show frequency in relative numbers}
}
\value{
list with descriptives.
}
\description{
Descriptives of nominal variables
}
