% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desc_perc}
\alias{desc_perc}
\title{Show frequency (relative) of answers of list of variables}
\usage{
desc_perc(dat, desc, valid = NULL, digits = 1)
}
\arguments{
\item{dat}{data.frame; contains sociodemographic and booklet variables}

\item{desc}{character vector; contains the name(s) of the variable(s) that
shall be used in analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{digits}{integer; number of decimals for rounding}
}
\value{
list with frequency (relative) of answers for each variable in desc.
}
\description{
Show frequency (relative) of answers of list of variables
}
