% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{describe}
\alias{describe}
\title{Calculate basic descriptive statistics for variables}
\usage{
describe(x, digits = 2)
}
\arguments{
\item{x}{A data frame with variables to calculate the statistics for}

\item{digits}{A number for rounding}
}
\value{
A data frame with the calculated statistics for each variable
}
\description{
Calculate basic descriptive statistics for variables
}
