% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{dif_summary}
\alias{dif_summary}
\title{Summary for DIF analysis}
\usage{
dif_summary(
  diflist,
  vars,
  print = TRUE,
  save = TRUE,
  path = "Tables",
  dif_threshold = 0.5,
  suf_item_names = FALSE,
  overwrite = FALSE,
  name_group = NULL,
  digits = 3L
)
}
\arguments{
\item{diflist}{list; return object of dif_model(); with main and dif model}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names}

\item{print}{logical; whether results shall be printed to console}

\item{save}{logical; whether results shall be saved to hard drive}

\item{path}{string; defines path to folder where tables shall be saved}

\item{dif_threshold}{numeric scalar; indicates absolute threshold of
problematic DIF (defaults to 0.5)}

\item{suf_item_names}{logical; whether to output SUF item names in the .xlsx file
for items with collapsed categories}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}

\item{digits}{integer; number of decimals for rounding}
}
\value{
list of information criteria, dif estimates and main effects in
data frames for dif analysis.
}
\description{
Summary for DIF analysis
}
