% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_analysis.R
\name{dimension_analysis}
\alias{dimension_analysis}
\title{Dimensionality analysis}
\usage{
dimension_analysis(
  resp,
  vars,
  items,
  dim = NULL,
  valid = NULL,
  irtmodel = "PCM2",
  maxiter = 10,
  snodes = 5,
  verbose = FALSE
)
}
\arguments{
\item{resp}{data.frame. contains:
(1) the responses. y in {0, 1} for binary data and y in
  {0, 1, ... k-1} for polytomous responses with k categories
(2) ID_t: column indicating ID of participants}

\item{vars}{data.frame with all variables as rows with at least the following columns:
items: contains name for each item (scored and unscored)
poly: identifies all polytomous (?) items (logical)
scored: identifies all scored items (logical)}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{dim}{character vector. contains names of all dimension variables that shall be analyzed}

\item{valid}{character string. defines name of boolean variable in dat,
indicating (in)valid cases.}

\item{irtmodel}{the choice of irtmodel as passed to the TAM function}

\item{maxiter}{max iterations as passed to the TAM function}

\item{snodes}{snodes as passed to the TAM function}

\item{verbose}{verbose as passed to the TAM function}
}
\value{
list of results for each dimensional analysis
}
\description{
Dimensionality analysis
}
