% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor_analysis.R
\name{distractor_analysis}
\alias{distractor_analysis}
\title{Distractor analysis}
\usage{
distractor_analysis(resp, vars, items, min.val = 3)
}
\arguments{
\item{resp}{data.frame with all respondents as rows and all variables as columns,
needs to contain at least all competence variables}

\item{vars}{data.frame with all variables as rows with at least following columns:
items: contains all item names (both scored and unscored)
scored: logical or binary integer, identifies all scored items
type: characer vector, identifies type of each item (e.g. "MC")
correct_response: integer, identifies correct response for each item}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{min.val}{minimum number of valid values; if negative,
set to the number of variables}
}
\value{
list of one data frame per item containing item-total correlations
  for each possible response; correct response is marked with an *
}
\description{
Distractor analysis
}
