% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_suf_and_sum_scores.R
\name{estimated_rotated_wles}
\alias{estimated_rotated_wles}
\title{Estimated WLEs with test rotation}
\usage{
estimated_rotated_wles(
  resp,
  vars,
  select,
  valid = NULL,
  facet,
  xsi_fixed,
  scoring,
  mvs,
  wle_name,
  irtmodel
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person
identifier and all variables that are further defined in the function
arguments; if special person sets are of interest, these persons have to
be selected in resp beforehand}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{facet}{character vector; contains the variable name indicating the
test rotation}

\item{xsi_fixed}{matrix; fixed parameter estimates of final IRT scaling as
returned by tam.mml()}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; can be NULL for
Rasch model}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{wle_name}{character; name of the wle -- WITHOUT extension (e.g.,
reg4 instead of reg4_sc1 or mag12 instead of mag12_sc1u)}

\item{irtmodel}{string; "1PL" for Rasch, or "PCM2" for PCM analysis}
}
\value{
a data.frame containing ID_t, wle and se of wle (named like indicated
  in wle_name)
}
\description{
Estimated WLEs with test rotation
}
