% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex3}
\alias{ex3}
\title{Polytomous NEPS data with three booklets}
\format{
A list with two data frames including the simulated responses
(\code{resp}) and the characteristics of the variables (\code{vars}).

\code{resp} is a data frame with 2,110 rows and 46 columns:
\describe{
\item{ID_t}{Unique identifier for respondent.}
\item{sex}{Sex with 0 = male and 1 = female.}
\item{mig}{Migrant background with 0 = no and 1 = yes.}
\item{rotation}{Test position with 0 = first and 1 = second.}
\item{booklet}{Booklet with 0 = easy, 1 = medium, and 2 = difficult.}
\item{reg7XXX_c}{Scored item responses 0 = incorrect and 1 = correct.}
\item{reg7XXX}{Unscored item responses for simple multiple-choice items
with values of 1 to 4 representing the selected response option.}
\item{valid}{A logical vector identifying valid respondents.}
\item{booklet1}{A logical identifying respondents for Booklet 1.}
\item{booklet2}{A logical identifying respondents for Booklet 2.}
\item{booklet3}{A logical identifying respondents for Booklet 3.}
}

\code{vars} is a data frame with 37 rows and 16 columns:
\describe{
\item{item}{Name of the item.}
\item{raw}{A logical distinguishing scored (\code{FALSE}) and unscored (\code{TRUE})
items.}
\item{dich}{A logical identifying scored dichotomous (\code{TRUE}) and
other (\code{FALSE}) items.}
\item{poly}{A logical identifying scored polytomous (\code{TRUE}) and other
(\code{FALSE}) items.}
\item{mixed}{A logical identifying scored (\code{TRUE}) and other (\code{FALSE})
items.}
\item{type}{A string identifying the item type as simple multiple-choice
(\code{MC}) or complex multiple-choice (\code{CMC}).}
\item{texttype}{The five text types included in the reading test.}
\item{scoring}{A number with the scoring weight in the item response
analyses.}
\item{num_cat}{A number with the number of response categories.}
\item{booklet1}{A logical identifying items included in Booklet 1.}
\item{pos1}{A number with the item position in Booklet 1.}
\item{booklet2}{A logical identifying items included in Booklet 2.}
\item{pos2}{A number with the item position in Booklet 2.}
\item{booklet2}{A logical identifying items included in Booklet 3.}
\item{pos3}{A number with the item position in Booklet 3.}
\item{correct}{A number between 1 and 4 representing the correct response
option for simple multiple-choice items.}
}
}
\usage{
ex3
}
\description{
The simulated data resembles the NEPS data structure for a test with 16
dichotomous and 5 polytomous items presented in three booklets.
}
\keyword{datasets}
