% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_figures.r
\name{Fig}
\alias{Fig}
\title{Generate figures in NEPS Survey Papers}
\usage{
Fig(
  file,
  width = NULL,
  height = NULL,
  crop = NULL,
  dpi = 96,
  footnote = NULL,
  note.offset = NULL,
  note.bgcol = "white",
  note.col = "black",
  note.x = 30,
  note.y = 70,
  note.size = 14
)
}
\arguments{
\item{file}{The file path for the image.}

\item{width}{The width of the resized image.}

\item{height}{The height of the resized image}

\item{crop}{A string indicating how to crop the image in the format
\code{"<width>x<height>{+-}<xoffset>{+-}<yoffset>"}
(see http://www.imagemagick.org/Magick++/Geometry.html)}

\item{dpi}{The dots per inch.}

\item{footnote}{An optional note to be displayed below the figure.}

\item{note.offset}{The height of the additional field below  the original
image that will include the note.}

\item{note.bgcol}{The background color of the note field.}

\item{note.col}{The font color of the note.}

\item{note.x}{The left position of the note.}

\item{note.y}{The top position of the note.}

\item{note.size}{The font size of the note.}
}
\value{
An image.
}
\description{
Load image from file, add note, and print it
}
