% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_figures.r
\name{FigMv}
\alias{FigMv}
\alias{FigMV}
\title{Crop plot for missing values to remove title}
\usage{
FigMv(file, footnote = NULL)

FigMV(file, footnote = NULL)
}
\arguments{
\item{file}{A file path for the image.}

\item{footnote}{An additional note to be displayed below the figure.}
}
\value{
An image.
}
\description{
Crop plot for missing values to remove title
}
