% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetDif}
\alias{GetDif}
\alias{GetDIF}
\title{Get values from DIF analyses}
\usage{
GetDif(
  obj,
  n = NULL,
  main = NULL,
  dif = NULL,
  item = FALSE,
  group = NULL,
  model = "dif",
  signed = FALSE,
  digits = 2
)

GetDIF(
  obj,
  n = NULL,
  main = NULL,
  dif = NULL,
  item = FALSE,
  group = NULL,
  model = "dif",
  signed = FALSE,
  digits = 2
)
}
\arguments{
\item{obj}{A data frame with sheets from "dif_dich_XXX.xlsx" or
"dif_poly_XXX.xlsx" created by \code{\link[=dif_analysis]{dif_analysis()}}.}

\item{n}{A group name to return the sample size of this group
(corresponds to the values in column "number" of sheet "facet").}

\item{main}{Return standardized main effects (\code{std}) or unstandardized
main effects (\code{ustd}). If the DIF variable has more than two groups
and \code{group} is missing, results for the first group comparison are returned.}

\item{dif}{A function used to summarize the DIF effects or a string
specifying an operator (=, <, >) with a number; e.g.,
\itemize{
\item =1 returns number of DIF effects equal to 1
\item <1 returns number of DIF effects less than 1
\item 1 returns number of DIF effects greater than 1Multiple conditions can be combined using & and |.If the DIF variable has more than two group and group is missing, resultsacross all groups are returned.
}}

\item{item}{A logical to return the item names (\code{TRUE}) instead of the
calculated statistic for the DIF effects (\code{FALSE}).}

\item{group}{If the DIF variable has more than two groups, the name of
the group comparison.}

\item{model}{The name of the model for which the main effects should be
returned, one of \code{dif} or \code{main}.}

\item{signed}{A logical to use the unsigned DIF effects (\code{FALSE}) instead of
the signed DIF effects (\code{TRUE}).}

\item{digits}{A number for rounding.}
}
\value{
The calculated DIF statistics or a vector of item names.
}
\description{
Get values from DIF analyses
}
\examples{
\dontrun{
data("ex1")

# DIF analyses
tmpdir <- tempdir()
dif_analysis(
 resp = ex1$resp,
 vars = ex1$vars,
 select = "dich",
 valid = "valid",
 dif_vars = c("sex", "mig"),
 include_mv = 100,
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 verbose = FALSE,
 warn = FALSE
)

# Import results of DIF analyses
dif <- Import(tmpdir, regexp = "^dif_dich_([^_]+\\\\.xlsx)")

# Sample size for men (coded as 0)
GetDif(dif$sex, n = 0)

# Standardized main effect in the DIF model for sex
GetDif(dif$sex, main = "std", model = "dif")

# Unstandardized main effect in the main effect model for mig
# for the comparison of migrants (coded 1) and the missing group (coded 3)
GetDif(dif$mig, main = "ustd", group = "1-3", model = "main")

# Median of the unsigned (absolute) DIF effects for sex
GetDif(dif$sex, dif = median)

# Largest signed DIF effects for mig for the comparision of migrants
# (coded 1) and non-migrants (coded 2)
GetDif(dif$mig, dif = max, signed = TRUE, group = "1-2")

# Number of unsigned (absolute) DIF effects for mig greater than .4
# across all groups
GetDif(dif$mig, dif = ">.4")

# Items with signed DIF effects for mig greater than .4
# across all groups
GetDif(dif$mig, dif = ">.4", signed = TRUE, item = TRUE)

# Clean up generated files
file.remove(paste0(tmpdir, "/dif_dich_sex.xlsx"))
file.remove(paste0(tmpdir, "/dif_dich_mig.xlsx"))
file.remove(paste0(tmpdir, "/dif_dich_TR.xlsx"))
file.remove(paste0(tmpdir, "/dif_dich_models.rds"))
file.remove(paste0(tmpdir, "/dif_dich_summaries.rds"))
}
}
