% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetDimFit}
\alias{GetDimFit}
\title{Get model fit for dimensionality analyses}
\usage{
GetDimFit(obj, model, type)
}
\arguments{
\item{obj}{A data frame with sheets from "dimensionality.xlsx"
created by \code{\link[=dim_analysis]{dim_analysis()}}.}

\item{model}{A column name identifying the model in \code{obj} which is typically
one of uni or dim.}

\item{type}{A row name identifying the model fit statistic in \code{obj} which
is typically one of Npars, loglik, AIC, or BIC.}
}
\value{
The model fit statistic.
}
\description{
Get model fit for dimensionality analyses
}
\examples{
\dontrun{
data("ex2")

# Dimensionality analyses
tmpdir <- tempdir()
ex2$vars$content <- as.numeric(ex2$vars$content)
dim_analysis(
 resp = ex2$resp,
 vars = ex2$vars,
 select = "mixed",
 valid = "valid",
 dim = "content",
 snodes = 5,# only for speed; model will not converge!
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 verbose = FALSE,
)

# Import results of dimensionality analyses
dim <- Import(tmpdir, "/dimensionality.xlsx")

# AIC for multidimensonal model for variable content
GetDimFit(dim, "content", "AIC")

# BIC for unidimensional model
GetDimFit(dim, "uni", "BIC")

# Clean up generated files
file.remove(paste0(tmpdir, "/dimensionality.xlsx"))
file.remove(paste0(tmpdir, "/dimensionality.rds"))
}
}
