% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetDist}
\alias{GetDist}
\title{Summarize distractor analyses}
\usage{
GetDist(obj, stat = median, correct = FALSE, item = FALSE, digits = 2)
}
\arguments{
\item{obj}{A data frame with sheets from "distractors_summary.xlsx"
created by \code{\link[=dis_analysis]{dis_analysis()}}.}

\item{stat}{A function used to summarize the distractor correlations.}

\item{correct}{A logical indicating whether to return the results for
distractors (\code{FALSE}) or correct response options (\code{TRUE}).}

\item{item}{A logical indicating whether to return the results from \code{stat}
(\code{FALSE}) or the item name corresponding to the value of \code{stat} (\code{TRUE}).}

\item{digits}{A number for rounding.}
}
\value{
The calculated statistic or a vector of item names.
}
\description{
Summarize distractor analyses
}
\examples{
data("ex2")

# Distractor analyses
tmpdir <- tempdir()
dis_analysis(
 resp = ex1$resp,
 vars = ex1$vars,
 select_raw = "raw",
 correct = "correct",
 valid = "valid",
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 warn = FALSE
)

# Import results of distractor analyses
dist <- Import(tmpdir, "distractors_summary.xlsx")

# Median total-rest correlations for the distractors
GetDist(dist, median)

# Smallest total-rest correlations for the correct responses
GetDist(dist, min, correct = TRUE)

# Item with the largest total-rest correlations for the distractors
GetDist(dist, max, item = TRUE)

# Clean up generated files
file.remove(paste0(tmpdir, "/distractors_summary.xlsx"))
file.remove(paste0(tmpdir, "/distractors_items.xlsx"))
file.remove(paste0(tmpdir, "/distractors.rds"))
}
