% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetFit}
\alias{GetFit}
\title{Get model fit}
\usage{
GetFit(obj, type, GPCM = FALSE, digits = 2)
}
\arguments{
\item{obj}{A data frame with sheets from "irt_dich.xlsx" or "irt_poly.xlsx"
created by \code{\link[=irt_analysis]{irt_analysis()}}.}

\item{type}{A column name in \code{obj} identifying the model fit statistic.}

\item{GPCM}{A logical to return the GPCM fit (\code{TRUE}) instead of the PCM fit
(\code{FALSE}).}

\item{digits}{A number for rounding.}
}
\value{
The model fit statistic.
}
\description{
Get model fit
}
\examples{
\dontrun{
data("ex1")

# IRT analyses
tmpdir <- tempdir()
irt_analysis(
 resp = ex1$resp,
 vars = ex1$vars,
 select = "dich",
 valid = "valid",
 plots = FALSE,
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 warn = FALSE,
 verbose = FALSE,
)

# Import results of IRT analyses
pars <- Import(tmpdir, "irt_dich.xlsx")

# BIC for the Rasch model
GetFit(pars, "BIC")

# Number of parameters in the 2PL
GetFit(pars, "Npars", GPCM = TRUE)

# Clean up generated files
file.remove(paste0(tmpdir, "/irt_dich.xlsx"))
file.remove(paste0(tmpdir, "/ir.rds"))
}
}
