% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetMvi}
\alias{GetMvi}
\alias{GetMVI}
\title{Calculate statistics for missing values by item}
\usage{
GetMvi(obj, type, stat, digits = 0)

GetMVI(obj, type, stat, digits = 0)
}
\arguments{
\item{obj}{A list with data frames with sheets from "mv_item.xlsx"
or a single data frame with the sheet "summary" from "mv_item.xlsx" created
by \code{\link[=mv_item]{mv_item()}}.}

\item{type}{The type of missing values identifying the row in \code{obj} which is
typcially one of OM, NR, NV, ND, or ALL.}

\item{stat}{The calculated statistic identifying the column in \code{obj} which is
typically one of Mean, SD, Median, Min, or Max.}

\item{digits}{A number for rounding.}
}
\value{
A number with the calculated statistic.
}
\description{
Calculate statistics for missing values by item
}
\examples{
data("ex1")

# Missing analyses by item
tmpdir <- tempdir()
mv_item(
 resp = ex1$resp,
 vars = ex1$vars,
 select = "dich",
 valid = "valid",
 pos = "pos",
 plots = FALSE,
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 warn = FALSE,
 verbose = FALSE,
)

# Import results of missing analyses
mvi <- Import(tmpdir, "/mv_item.xlsx")

# Largest percentage of invalid responses
GetMvi(mvi, "NV", "Max")

# Median percentage of missing values with two decimals
GetMvi(mvi, "ALL", "Median", digits = 2)

# Clean up generated files
file.remove(paste0(tmpdir, "/mv_item.xlsx"))
file.remove(paste0(tmpdir, "/mv_item.rds"))
}
