% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetProp}
\alias{GetProp}
\title{Extractor functions for NEPS Survey Papers}
\usage{
GetProp(obj, select, prop = "type", val = c("CMC", "MA"), item = FALSE)
}
\arguments{
\item{obj}{A data frame with item characteristics.}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{prop}{A variable name in \code{vars} identifying the item property such
as response format, text function, or cognitive function (default: \code{type}).}

\item{val}{The values of \code{prop} to count.}

\item{item}{A logical to return the item names (\code{TRUE}) or the numeric
values (\code{FALSE}).}
}
\value{
The number of items or the item names.
}
\description{
Determine the number of items with a given item property
}
\examples{
data("ex2")

# Number of CMC items
GetProp(ex2$vars, select = "mixed", prop = "type", val = "CMC")

# Names of CMC items
GetProp(ex2$vars, select = "mixed", prop = "type", val = "CMC",
                 item = TRUE)
}
