% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetPropLabels}
\alias{GetPropLabels}
\title{Get long label for item property}
\usage{
GetPropLabels(short = NULL)
}
\arguments{
\item{short}{A short label}
}
\value{
A long label
}
\description{
Get long label for item property
}
