% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetVar}
\alias{GetVar}
\title{Get variance of latent proficiency distribution}
\usage{
GetVar(obj, digits = 2)
}
\arguments{
\item{obj}{A \code{\link[TAM:tam.mml]{TAM::tam.mml()}} object with a fitted IRT model in
"irt_dich.rds" or "irt_poly.rds" generated by \code{\link[=irt_analysis]{irt_analysis()}}.}

\item{digits}{A number for rounding.}
}
\value{
The population variance.
}
\description{
Get variance of latent proficiency distribution
}
\examples{
\dontrun{
data("ex1")

# IRT analyses
tmpdir <- tempdir()
irt_analysis(
 resp = ex1$resp,
 vars = ex1$vars,
 select = "dich",
 valid = "valid",
 plots = FALSE,
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 warn = FALSE,
 verbose = FALSE,
)

# Import results of IRT analyses
irt <- readRDS(paste0(tmpdir, "/irt_dich.rds"))

# Population variance
GetVar(irt)

# Clean up generated files
file.remove(paste0(tmpdir, "/irt_dich.xlsx"))
file.remove(paste0(tmpdir, "/irt_dich.rds"))
}
}
