% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{grouped_irt_analysis}
\alias{grouped_irt_analysis}
\title{IRT analyses for several groups}
\usage{
grouped_irt_analysis(
  groups,
  resp,
  vars,
  valid = NULL,
  mvs = NULL,
  scoring = NULL,
  plots = FALSE,
  save = TRUE,
  print = TRUE,
  return = FALSE,
  overwrite = FALSE,
  path_plots = here::here("Plots"),
  path_table = here::here("Tables"),
  path_results = here::here("Results"),
  digits = 3,
  verbose = FALSE,
  warn = TRUE,
  xsi.fixed = NULL,
  pweights = NULL,
  control_tam = NULL,
  control_wle = NULL
)
}
\arguments{
\item{groups}{named character vector; contains name of item selection
variable of each group (e.g. easy = 'easy_final')}

\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; can be NULL for
Rasch model}

\item{plots}{logical; whether plots shall be created and saved to hard drive}

\item{save}{logical; whether results shall be saved to hard drive}

\item{print}{logical; whether results shall be printed to console}

\item{return}{logical; whether results shall be returned}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{path_plots}{string; defines path to folder where plots shall be saved}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{path_results}{string; defines path to folder where results shall be saved}

\item{digits}{integer; number of decimals for rounding}

\item{verbose}{logical; whether to print processing information to console}

\item{warn}{logical; whether to print warnings (should be set to TRUE to
avoid problems with the data structure)}

\item{xsi.fixed}{matrix; contains fixed (linked) item parameters; optional}

\item{pweights}{character; defines name of numerical variable in resp that
contains person weights as passed to TAM-function}

\item{control_tam}{list; control argument as passed to tam.mml-functions}

\item{control_wle}{list; can contain Msteps and/or convM as to pass to tam.wle()
as elements of the list}
}
\value{
list with a list for each group with all results
(as return object of irt_analysis())
}
\description{
IRT analyses for several groups
}
