% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{hl_scoring}
\alias{hl_scoring}
\title{Score highlighting items}
\usage{
hl_scoring(
  resp,
  hl_solutions,
  hl_distractors,
  mvs = NULL,
  warn = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains original item responses}

\item{hl_solutions}{list; contains character vector with subitems for each
highlighting item with correct solutions, name of the vector is the name of the highlighting item (e.g.
hl_solutions = list(hl1 = c("subitem1", "subitem2"), hl2 = c("subitem1", "subitem2"))),
name of the vector should match the vector names in @hl_distractors}

\item{hl_distractors}{list; contains character vector with subitems for each
highlighting item with distractors, name of the vector is the name of the highlighting item (e.g.
hl_distractors = list(hl1 = c("subitem1", "subitem2"), hl2 = c("subitem1", "subitem2"))),
name of the vector should match the vector names in @hl_solutions}

\item{mvs}{integer vector; contains user-defined missing values}

\item{warn}{logical; print warnings}

\item{verbose}{logical; print messages}
}
\value{
resp including unscored (raw) and scored items
}
\description{
Score highlighting items
}
