% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{icc_plots}
\alias{icc_plots}
\title{ICC plots}
\usage{
icc_plots(results, name, path = here::here("Plots"))
}
\arguments{
\item{results}{list. Contains results from IRT analysis as returned by function
'irt_analysis'.}

\item{name}{character. Part of the filename for the directory of the plots,
following "ICCs_for_" (e.g., the kind of IRT analysis: "1PL").}

\item{path}{character. contains name of path for plots.}
}
\description{
Create ICC plots for IRT models.
}
