% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_import.r
\name{Import}
\alias{Import}
\title{Import an excel file with all sheets}
\usage{
Import(
  path,
  filename = NULL,
  sheet = NULL,
  regexp = NULL,
  rename = NULL,
  replace = c(dif_poly_ = "", dif_dich_ = "", mv_person_ = "")
)
}
\arguments{
\item{path}{A path to the folder with the result tables created by the
\code{NEPSroutines} package.}

\item{filename}{An optional name of an excel file to import or a vector of
multiple file names; if not set \code{regexp} should be used.}

\item{sheet}{An optional name of an excel sheet; if set, only the specified
excel sheet will be imported.}

\item{regexp}{An optional regular expression to identify file names in the
folder given by \code{path}; if not set, \code{filename} should be used.}

\item{rename}{An optional named vector with new names for the imported files;
the names of the vector need to identify the original file names.}

\item{replace}{An optional named vector with substrings to replace in the
names of the imported files; the names identify the original string, while
the value gives the replacement.}
}
\value{
A list of data frames if a multiple files were imported or a single
data frame if a single file was imported.
}
\description{
Import an excel file with all sheets
}
\examples{
\dontrun{
# Import all sheets from file 'mv_item.xlsx' located in the folder 'tables'
mvi <- Import(path = "tables", filename = "mv_item.xlsx")

# Import sheet 'summary' from file 'mv_item.xlsx' located in the folder 'tables'
mvisum <- Import(path = "tables", filename = "mv_item.xlsx",
                 sheet = "summary")

# Import all sheets from files with names that begin with "dif_poly_" and
# end with "_isRegular" located in the folder 'tables'; the file named
# "dif_poly_TR_isRegular" is renamed to "TR" in the returned data.frame.
mvi <- Import(path = "tables",
              regexp = "dif_poly_(.+)_isRegular"),
              rename = c("dif_poly_TR_isRegular" = "TR"))
}
}
