% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{irt_all}
\alias{irt_all}
\title{IRT analyses - all in one function}
\usage{
irt_all(
  resp,
  vars,
  items,
  irt_type,
  Q = NULL,
  icc_plots = TRUE,
  wright_map = TRUE,
  path_plots = here::here("Plots"),
  name_table = NULL,
  path_table = here::here("Tables"),
  digits = 2,
  return_results = TRUE,
  name_data = NULL,
  path_data = here::here("Data")
)
}
\arguments{
\item{resp}{data.frame. contains:
(1) the responses. y in {0, 1} for binary data and y in
  {0, 1, ... k-1} for polytomous responses with k categories
(2) ID_t: column indicating ID of participants}

\item{vars}{data.frame. contains all competence items as rows,
and at least the following variables:
  character vector named "items"; contains the names of the items.
  boolean vector; indicates which items to use for analysis.}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{Q}{matrix with one column and ncol(resp) rows. Binary items are
scored with 1, polytomous items are scored with 0.5.
Divergent scoring decisions can be incorporated by this matrix
(e.g., item 7 is scored with 0.75 by assigning 0.75 to the seventh row).}

\item{digits}{number of decimals for rounding}

\item{return_results}{boolean. indicates whether to return results.}

\item{irtmodel}{character. "1PL" for Rasch, "2PL" for 2PL, "PCM2" for PCM and
"GPCM" for GPCM analyses.}

\item{path}{folder path for data}

\item{filename}{string with name of file that shall be saved (including file type).}

\item{verbose}{logical. If verbose == TRUE information about the estimation
progress is printed to the console}
}
\description{
Perform 1PL and 2PL analyses for binary or PCM and GPCM analyses for
polytomous data, generate summary and plots.
}
