% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{irt_model_fit}
\alias{irt_model_fit}
\title{IRT model fit}
\usage{
irt_model_fit(
  model_dich,
  model_poly,
  irt_type,
  overwrite = FALSE,
  path = here::here("Tables"),
  filename = NULL
)
}
\arguments{
\item{model_dich}{list; return object of irt_model(); dichotomous analysis}

\item{model_poly}{list; return object of irt_model(); polytomous analysis}

\item{irt_type}{string; either "dich" (dichotomous analysis) or "poly"
(polytomous analysis)}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{path}{string; defines path to folder where table shall be saved}

\item{filename}{string; defines name of table that shall be saved}
}
\value{
data.frame with AIC, BIC and number of parameters for both models
}
\description{
Create table with model fit for 1 parameter and 2 parameter IRT model.
}
