% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{link_wles}
\alias{link_wles}
\title{Link WLEs using a known linking constant while correcting for selection}
\usage{
link_wles(
  wle_prev,
  wle_curr,
  const,
  wid = "wle",
  use_longitudinal_subsample = TRUE
)
}
\arguments{
\item{wle_prev}{data.frame with WLEs at first measurement wave}

\item{wle_curr}{data.frame with WLEs at second measurement wave}

\item{const}{linking constant}

\item{wid}{character; variable name used as WLE identifier}

\item{use_longitudinal_subsample}{boolean indicating whether linking should
be based on the entire samples given in wle_prev and wle_curr or
only on the longitudinal sample with common pids; defaults to TRUE}
}
\value{
linked WLEs at second measurement wave
}
\description{
Link WLEs using a known linking constant while correcting for selection
}
