% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meht}
\alias{meht}
\title{Minimum effect hypothesis test
(following Murphy & Myors, 1999)}
\usage{
meht(
  stat,
  df1,
  df2,
  eta2 = NULL,
  delta = 0.4,
  alpha = 0.05,
  digits = 3,
  verbose = TRUE
)
}
\arguments{
\item{stat}{empirical F statistic}

\item{df1}{degrees of freedom (= number of groups - 1)}

\item{df2}{degrees of freedom (= sample size - number of groups)}

\item{eta2}{percentage of explained variance (= minimum effect size)}

\item{delta}{standardized mean difference (= minimum effect size)}

\item{alpha}{error probability}

\item{digits}{number of decimal places}

\item{verbose}{print results to console}
}
\value{
list
}
\description{
Minimum effect hypothesis test
(following Murphy & Myors, 1999)
}
