% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_item.R
\name{mvi_plots}
\alias{mvi_plots}
\title{Plot percentages of user-defined missing values by item position}
\usage{
mvi_plots(
  vars,
  items,
  mv_i = NULL,
  resp = NULL,
  valid = NULL,
  position = NULL,
  grouping = NULL,
  show_all = FALSE,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  labels_mvs = c(ALL = "total missing items", OM = "omitted items", NV =
    "not valid items", NR = "not reached items", TA =
    "missing items due to test abortion", UM = "unspecific missing items", ND =
    "not determinable items", NAd = "not administered items", AZ =
    "missing items due to ''Angabe zurueckgesetzt''"),
  path = here::here("Plots/Missing_Responses/by_item"),
  color = NULL,
  digits = 2,
  verbose = TRUE,
  warn = TRUE
)
}
\arguments{
\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{items}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{mv_i}{list; return object of mvi_analysis()}

\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; all responses ∈ ℕ0; user-defined missing values;
includes all variables that are further defined in the function arguments}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{position}{(named) character vector; defines name(s) of integer
variable(s) in vars that indicate position of items; if groups with differing
item positions in testlets exist, then vector must be named with names of
groups (as in "grouping") as names of elements and names of variables as elements}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{show_all}{logical; whether whole sample shall be included as a "group"
(only applicable when grouping exists)}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{labels_mvs}{named character vector; contains labels for user-defined
missing values to use them in plot titles and printed results}

\item{path}{string; defines path to folder where plots shall be saved}

\item{color}{character calar or vector; defines color(s) of the bar in plots}

\item{digits}{integer; number of decimals for rounding}

\item{verbose}{logical; whether to print processing information to console}

\item{warn}{logical; whether to print a warning if NAs were found in resp}
}
\description{
Plots the percentages of not reached missing values for each item by its
item position. If not otherwise specified, the order in which the items are
given to the function is taken as the order the items appeared in the test.
}
