% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_item.R
\name{mvi_table}
\alias{mvi_table}
\title{Write table of user defined missing values per item (and position)}
\usage{
mvi_table(
  vars,
  items,
  mv_i = NULL,
  resp = NULL,
  position = NULL,
  grouping = NULL,
  show.all = FALSE,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  filename = NULL,
  path = here::here("Tables"),
  min.val = 3,
  digits = 2,
  warn = TRUE,
  return_table = FALSE
)
}
\arguments{
\item{vars}{data.frame. contains all competence items as rows,
and at least the following variables:
character vector named "items"; contains the names of the items.
boolean vector; indicates which items to use for analysis.}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{mv_i}{list; return object of mvi_analysis(). mvi_analysis$list must contain
for each group and all groups together the name of the
user-defined missing values that are to be
analyzed and the names of the items as its row names.}

\item{resp}{data.frame. contains item responses (integer with user-defined missing values)
and grouping variables (boolean).}

\item{position}{(named) character vector. contains name(s) of variable(s) in
vars that indicate the position of subitems;
if grouping with differing positions in testlets,
then for each group one variable name is necessary,
as well as names for the vector that represent names of groups
and must be identical with items in variable "grouping"
(e.g. position = c(easy = "position_easy", diff = "position_diff"))}

\item{grouping}{character vector. contains names of groups (e.g. 'easy and 'difficult')
as used in resp.}

\item{show.all}{boolean; only needed when groups exist, indicates whether
plots shall also include the whole sample as a "group"}

\item{mvs}{named vector with definition of user-defined missing values.}

\item{filename}{string with name of file that shall be saved (including type of file);
if NULL, no file will be saved.}

\item{path}{folder path for plots}

\item{min.val}{minimum number of valid values;
if negative, set to the default of 3}

\item{digits}{number of decimals for rounding}

\item{warn}{boolean whether to print a warning if NAs were found in resp}

\item{return_table}{boolean. indicates whether to return table.}
}
\description{
Plots the percentages of not reached missing values for each item by its
item position. If not otherwise specified, the order in which the items are
given to the function is taken as the order the items appeared in the test.
}
