% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{mvp_analysis}
\alias{mvp_analysis}
\title{Percentage of missing responses by person}
\usage{
mvp_analysis(
  resp,
  vars,
  items,
  valid = NULL,
  grouping = NULL,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  filename = NULL,
  path = here::here("Results"),
  digits = 2,
  warn = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; all responses ∈ ℕ0; user-defined missing values;
includes all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{items}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{filename}{string; defines name of file that shall be saved}

\item{path}{string; defines path to folder where results shall be saved}

\item{digits}{integer; number of decimals for rounding}

\item{warn}{logical; whether to print a warning if NAs were found in resp}
}
\value{
list with missing values per person and
for each missing value type (in percentage) is returned
}
\description{
Percentage of missing responses by person
}
