% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{mvp_plots}
\alias{mvp_plots}
\title{Plot percentage of missing values for persons}
\usage{
mvp_plots(
  mv_p = NULL,
  resp = NULL,
  vars = NULL,
  items = "final",
  valid = NULL,
  grouping = NULL,
  show.all = FALSE,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  path = here::here("Plots/Missing_Responses/by_person"),
  filename = "Missing_responses_by_person",
  lbls = c(ALL = "total missing", OM = "omitted", NV = "not valid", NR = "not reached",
    TA = "test aborted", UM = "unspecific missing", ND = "not determinable", NAd =
    "not administered", AZ = "Angabe zurueckgesetzt"),
  color = NULL,
  verbose = TRUE
)
}
\arguments{
\item{mv_p}{percentage of missing responses by person as
returned by mv_person()}

\item{resp}{data.frame with item responses (user-defined missing values)
and logical grouping variables, named after the grouping
factor (e.g. "easy" and "difficult" in the case of testlet
difficulties) --> must be in separate columns!}

\item{vars}{data.frame. contains all competence items as rows,
with the column "items" indicating the names of teh items.}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to items in analysis.}

\item{valid}{character string. defines name of boolean variable in resp,
indicating (in)valid cases.}

\item{grouping}{character vector. contains names of groups (e.g. 'easy and 'difficult')
as itemsd in data.frame resp.}

\item{show.all}{logical; only needed when groups exist, indicates whether
plots shall also include the whole sample as a "group"}

\item{mvs}{named vector with definition of user-defined missing values}

\item{path}{folder path for plots}

\item{filename}{string with name of file that shall be saved (excluding type of file)}

\item{lbls}{lables for different types of user-defined missing values}

\item{color}{character vector. defines bar color(s); if the data contains subgroups,
one color per subgroup must be specified}

\item{verbose}{print progress}
}
\description{
Plot percentage of missing values for persons
}
