% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{mvp_table}
\alias{mvp_table}
\title{Write table of user defined missing values per item (and position)}
\usage{
mvp_table(
  mv_p,
  grouping = NULL,
  overwrite = FALSE,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, MD = -54, AZ = -21),
  missing_by_design = -54,
  save = TRUE,
  path = here::here("Tables"),
  name_group = NULL,
  test = TRUE
)
}
\arguments{
\item{mv_p}{list; return object of mv_person()}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{overwrite}{boolean; indicates whether to overwrite existing file when saving table}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{missing_by_design}{integer; user defined missing value for missing by
design}

\item{save}{logical; whether results shall be saved to hard drive}

\item{path}{string; defines path to folder where table shall be saved}

\item{name_group}{string; defines name of group used in analysis (e.g. 'settingA')}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\value{
table with missing values per person and for each missing value type
(in percentage) as well as summary statistics for TR.
}
\description{
Plots the percentages of not reached missing values for each item by its
item position. If not otherwise specified, the order in which the items are
given to the function is taken as the order the items appeared in the test.
}
