% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{n_valid}
\alias{n_valid}
\title{Number of valid cases}
\usage{
n_valid(dat, valid = NULL)
}
\arguments{
\item{dat}{data.frame with item responses (user-defined missing values)
and logical grouping variables, named after the grouping
factor (e.g. "easy" and "difficult" in the case of testlet.}

\item{valid}{character string. defines name of boolean variable in dat,
indicating (in)valid cases.}
}
\value{
character string with answer.
}
\description{
Number of valid cases
}
