% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{only_valid}
\alias{only_valid}
\title{Select only valid cases}
\usage{
only_valid(resp, valid = NULL, warn = TRUE)
}
\arguments{
\item{resp}{data.frame with item responses}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{warn}{logical; whether to warn if no parameter 'valid' provided}
}
\value{
resp without invalid cases.
}
\description{
Select only valid cases
}
