% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{prepare_longitudinal_resp}
\alias{prepare_longitudinal_resp}
\title{Prepare response data for linking}
\usage{
prepare_longitudinal_resp(
  resp_curr,
  resp_prev,
  resp_link = NULL,
  vars_curr,
  vars_prev,
  vars_link = NULL,
  select_curr,
  select_prev,
  select_link = NULL,
  valid_curr = NULL,
  valid_prev = NULL,
  valid_link = NULL,
  anchors = NULL,
  mvs = NULL,
  wid = "wle",
  use_longitudinal_subsample = TRUE,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{resp_curr}{data.frame with responses of current measurement wave,
a person identifier 'ID_t' and a variable indicating valid cases; may also
include pweights}

\item{resp_prev}{data.frame with responses of previous measurement wave,
a person identifier 'ID_t' and a variable indicating valid cases; may also
include a WLE as given in wid and pweights}

\item{resp_link}{data.frame with responses of link study, a person identifier
'ID_t' and a variable indicating valid cases; may also include pweights;
required if anchor group design is used}

\item{vars_curr}{data.frame; contains information about items of current
measurement wave with items as rows; includes variable 'item' containing item
names and may include variable with item scoring}

\item{vars_prev}{data.frame; contains information about items of previous
measurement wave with items as rows; includes variable 'item' containing item
names and may include variable with item scoring}

\item{vars_link}{data.frame; contains information about items of link study
with items as rows; includes variable 'item' containing item names and may
include variable with item scoring; required if anchor group design is used}

\item{select_curr}{character; contains name of logical variable in vars_curr
identifying the item set for the current measurement wave}

\item{select_prev}{character; contains name of logical variable in vars_prev
identifying the item set for the previous measurement wave}

\item{select_link}{character; contains name of logical variable in vars_link
identifying the item set for the link study; required if anchor group design
is used}

\item{valid_curr}{string; defines name of logical variable in resp_curr that
indicates (in)valid cases for the current measurement wave}

\item{valid_prev}{string; defines name of logical variable in resp_prev that
indicates (in)valid cases for the previous measurement wave}

\item{valid_link}{string; defines name of logical variable in resp_link that
indicates (in)valid cases for the link study; optional variable for anchor
group design}

\item{anchors}{character; data.frame with two columns including the link
items; for anchor item designs the first column refers to the previous
measurement time point and the second column to the current measurement
time point; for anchor group designs the first column refers to the main
sample (for both time points) and the second column refers to the link
sample; if NULL, all common items are used}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{wid}{variable name used as WLE identifier in first measurement wave}

\item{use_longitudinal_subsample}{logical; whether an intra-cohort link (TRUE)
using only persons taking part in both time points or
an inter-cohort link (FALSE) is conducted}

\item{warn}{logical; whether to print warnings}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\value{
list with response data
}
\description{
Prepare response data for linking
}
