% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_resp}
\alias{prepare_resp}
\title{Prepare resp for analysis}
\usage{
prepare_resp(
  resp,
  vars = NULL,
  items = NULL,
  use_only_valid = FALSE,
  valid = NULL,
  convert = FALSE,
  mvs = NULL,
  warn = TRUE,
  zap_labels = TRUE
)
}
\arguments{
\item{resp}{data.frame with item responses}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{items}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{use_only_valid}{logical; whether to check valid cases}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{convert}{logical; whether to convert custom missing values to NA}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{warn}{logical; whether warnings are to be printed to the console}

\item{zap_labels}{logical; whether to convert haven_labelled to normal}
}
\value{
data.frame as resp, but only with valid cases
}
\description{
Prepare resp for analysis
}
