% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{print_dif_summary}
\alias{print_dif_summary}
\title{Print DIF summaries to console}
\usage{
print_dif_summary(resp, diflist, res, dif_threshold = 0.5)
}
\arguments{
\item{resp}{data.frame; includes all DIF variables}

\item{diflist}{list; return object of dif_model(); with main and dif model}

\item{res}{list; return object of dif_summary()}

\item{dif_threshold}{numeric scalar; indicates absolute threshold of
problematic DIF (defaults to 0.5)}
}
\description{
Print DIF summaries to console
}
