% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_analysis.R
\name{print_dim_summary}
\alias{print_dim_summary}
\title{Print summary of dimensionality analysis}
\usage{
print_dim_summary(dimsum)
}
\arguments{
\item{dimsum}{return object of dim_summary()}
}
\description{
Print summary of dimensionality analysis
}
