% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor_analysis.R
\name{print_dis_summary}
\alias{print_dis_summary}
\title{Print summary of distractor analyses}
\usage{
print_dis_summary(dist_sum)
}
\arguments{
\item{dist_sum}{return object of dis_summary()}
}
\description{
Print summary of distractor analyses
}
