% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{print_irt_summary}
\alias{print_irt_summary}
\title{Print IRT results}
\usage{
print_irt_summary(model, irt_sum, steps_sum = NULL)
}
\arguments{
\item{model}{list; return object of irt_model()}

\item{irt_sum}{data.frame; return object of irt_summary()}

\item{steps_sum}{data.frame; return object of steps_analysis()}
}
\description{
Print and highlight IRT (and steps) analysis results.
}
