% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{print_link_results}
\alias{print_link_results}
\title{Print linking results for TR}
\usage{
print_link_results(link_dif, link_dim, link_results)
}
\arguments{
\item{link_dif}{return object of conduct_dif_anchor()}

\item{link_dim}{return object of check_link_dimensionality()}

\item{link_results}{return object of link_samples()}
}
\description{
Print linking results for TR
}
