% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_item.R
\name{print_mvi_results}
\alias{print_mvi_results}
\title{Print mvi results to console}
\usage{
print_mvi_results(
  mv_i,
  grouping = NULL,
  labels_mvs = c(ALL = "total missing items", OM = "omitted items", NV =
    "not valid items", NR = "not reached items", TA =
    "missing items due to test abortion", UM = "unspecific missing items", ND =
    "not determinable items", NAd = "not administered items", AZ =
    "missing items due to ''Angabe zurueckgesetzt''")
)
}
\arguments{
\item{mv_i}{list; return object of mvi_analysis()}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{labels_mvs}{named character vector; contains labels for user-defined
missing values to use them in plot titles and printed results}
}
\description{
Print mvi results to console
}
