% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{print_mvp_results}
\alias{print_mvp_results}
\title{Print mvp results to console}
\usage{
print_mvp_results(mv_p)
}
\arguments{
\item{mv_p}{list; return object of mvp_analysis()}
}
\description{
Print mvp results to console
}
