% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{props_by_version}
\alias{props_by_version}
\title{Table with item properties by test version}
\usage{
props_by_version(
  vars,
  select,
  grouping,
  properties,
  labels = NULL,
  save = FALSE,
  overwrite = FALSE,
  path = here::here("Tables"),
  warn = TRUE
)
}
\arguments{
\item{vars}{data.frame; contains information about the competence items}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{grouping}{character vector; contains for each group a name of a logical
variable in vars that indicates to which group belongs a person or an item}

\item{properties}{character vector; defines name(s) of variable(s) in vars
that identify item properties}

\item{labels}{list; may contain for each variable defined in argument 'properties'
a named character vector with the labels for the variable categories, the name
of the vector must be identical with the name identified in 'properties'
(e.g. sex = c(male = 1, female = 2))}

\item{save}{logical; whether to save the table in Excel}

\item{overwrite}{logical; whether to overwrite an existing table with the same name}

\item{path}{string; defines path for saving the table}

\item{warn}{logical; whether to print warnings (should be better set to true)}
}
\value{
table with item properties by test version
}
\description{
Table with item properties by test version
}
