% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recodeVar}
\alias{recodeVar}
\title{Recode variables}
\source{
doBy::recodeVar()
}
\usage{
recodeVar(x, src, tgt, default = NULL, keep.na = TRUE)
}
\arguments{
\item{x}{vector; the variable to be recoded}

\item{src}{source values: a subset of the present values of x}

\item{tgt}{target values: the corresponding new values of x}

\item{default}{default target value for those values of x not listed in
\code{src}. When \code{default=NULL}, values of x which are not given in
\code{src} will be kept in the output.}

\item{keep.na}{logical; if TRUE then NA's in x will be retained in the output}
}
\value{
vector
}
\description{
Recode variables
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
