% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rnd}
\alias{rnd}
\title{Rounding with proper formatting for NEPS Survey Papers}
\usage{
rnd(x, digits = 2, d0 = FALSE)
}
\arguments{
\item{x}{numeric vector; the numbers to be formatted}

\item{digits}{integer vector; the number of decimal places for rounding}

\item{d0}{logical vector; remove leading zeros (TRUE) or keep them (FALSE)}
}
\value{
character vector; the formatted numbers
}
\description{
Rounding with proper formatting for NEPS Survey Papers
}
