% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{sample_by_group}
\alias{sample_by_group}
\title{Table with sample size by groups}
\usage{
sample_by_group(
  resp,
  grouping_variable,
  labels = NULL,
  save = FALSE,
  overwrite = FALSE,
  path = here::here("Tables"),
  name_group = NULL
)
}
\arguments{
\item{resp}{data.frame; contains grouping variable}

\item{grouping_variable}{string; defines name of variable in resp that identifies groups}

\item{labels}{named character vector; links a label each value of groups
(e.g. labels = c(easy = 1, difficult = 2))}

\item{save}{logical; whether to save the table in Excel}

\item{overwrite}{logical; whether to overwrite an existing table with the same name}

\item{path}{string; defines path for saving the table}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}
}
\value{
data.frame with sample size by groups.
}
\description{
Table with sample size by groups
}
