% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{sample_by_version}
\alias{sample_by_version}
\title{Table with sample size by test version}
\usage{
sample_by_version(
  dat,
  versions,
  labels = NULL,
  save = FALSE,
  overwrite = FALSE,
  path = here::here("Tables")
)
}
\arguments{
\item{dat}{data.frame; contains version variable}

\item{versions}{string; defines name of variable in dat that identifies test versions}

\item{save}{logical; whether to save the table in Excel}

\item{overwrite}{logical; whether to overwrite an existing table with the same name}

\item{path}{string; defines path for saving the table}

\item{versions_lbls}{named character vector; links a label each value of versions
(e.g. versions_lbls = c(version1 = 1, version2 = 2))}
}
\value{
table with sample size by test version
}
\description{
Table with sample size by test version
}
