% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{save_dif_summary}
\alias{save_dif_summary}
\title{Save DIF summaries to hard disk}
\usage{
save_dif_summary(diflist, res, path_results, path_table, overwrite)
}
\arguments{
\item{diflist}{list; return object of dif_model(); with main and dif model}

\item{res}{list; return object of dif_summaries()}

\item{path_results}{string; defines path to folder where results shall be saved}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{overwrite}{logical; whether to overwrite existing file when saving table}
}
\description{
Save DIF summaries to hard disk
}
