% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_suf.R
\name{set_labels}
\alias{set_labels}
\title{Set labels of items in SUF}
\usage{
set_labels(
  suf,
  vars,
  select,
  competence,
  score_name = NULL,
  mvs = NULL,
  items_labels = NULL,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{suf}{data.frame; SUF to be (re)labelled}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{character; indicates the logical variable in vars which
contains the item names designated for the SUF (original scored items
before the first IRT analyses)}

\item{competence}{character; name of tested competence for labelling the
data}

\item{score_name}{character; name of the test scores -- WITHOUT extension (e.g.,
reg4 instead of reg4_sc1 or mag12 instead of mag12_sc1u)}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{items_labels}{vector; named vector for each item used for labelling
the data}

\item{warn}{logical; whether to print warnings}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\value{
suf with labels
}
\description{
Set labels of items in SUF
}
