% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_setup.r
\name{Setup}
\alias{Setup}
\title{Copy files for NEPS Survey Papers to local folder}
\usage{
Setup(
  path = getwd(),
  ask = TRUE,
  sc = 0,
  domain = "re",
  study = "A000",
  wave = 0,
  overwrite = FALSE
)
}
\arguments{
\item{path}{The path to the folder for the technical report.}

\item{ask}{A logical to ask for confirmation and some information (\code{TRUE})
or provide the information as arguments without confirmation (\code{FALSE}).}

\item{sc}{The number of the starting cohort.}

\item{domain}{A two letter abbreviation for the competence domain.}

\item{study}{The number of the study.}

\item{wave}{The number of the study wave.}

\item{overwrite}{A logical to overwrite existing files}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
Setup the folder with the Quarto extension for the technical report
}
\seealso{
\code{vignette("technical_report", package = "NEPSroutines")}
}
