% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_setup.r
\name{SetupRenderTemplate}
\alias{SetupRenderTemplate}
\title{Replace placeholders in template for Quarto report}
\usage{
SetupRenderTemplate(txt, data)
}
\arguments{
\item{txt}{The template with placeholders}

\item{data}{A list with data to be included in the template}
}
\value{
The template without placeholders
}
\description{
Replace placeholders in template for Quarto report
}
