% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{show_attributes}
\alias{show_attributes}
\title{Show attributes of selected variables}
\usage{
show_attributes(resp, desc)
}
\arguments{
\item{resp}{data.frame; contains sociodemographic and booklet variables}

\item{desc}{character vector; contains the name(s) of the variable(s) that
shall be used in analysis}
}
\description{
Show attributes of selected variables
}
