% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{steps_analysis}
\alias{steps_analysis}
\title{Step analysis}
\usage{
steps_analysis(
  results,
  path = here::here("Tables"),
  filename = NULL,
  digits = 2,
  return_table = TRUE
)
}
\arguments{
\item{results}{list. contains results from IRT analysis with one parameter
(PCM analysis).}

\item{path}{character. defines name of path for table.}

\item{filename}{character. defines name for excel document. if NULL (default),
the table will not be saved.}

\item{digits}{integer; how many digits after rounding}

\item{return_table}{logical; whether resulting table should be returned}
}
\value{
a data.frame containing the step parameters and SEs for each step
}
\description{
Create table with results of step analysis.
}
