% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{summarize_link_dif}
\alias{summarize_link_dif}
\title{Summarize link DIF results}
\usage{
summarize_link_dif(
  dif_summary = NULL,
  mod_current = NULL,
  mod_previous = NULL,
  items_previous,
  items_current,
  anchors = NULL,
  diff_threshold = 0.5,
  anchor_item_names = NULL
)
}
\arguments{
\item{dif_summary}{return object of dif_model() (anchor groups design)}

\item{mod_current}{return object of irt_model() for current assessment
(anchor items design)}

\item{mod_previous}{return object of irt_model() for previous assessment
(anchor items design)}

\item{items_previous}{character vector; item names of previous assessment}

\item{items_current}{character vector; item names of current assessment}

\item{anchors}{character vector with link items; if NULL, all common items
are used}

\item{diff_threshold}{numeric; threshold under which DIF in common link items
is accepted; defaults to .5}

\item{anchor_item_names}{character matrix; contains the old (column 1) and
new (column 2) item names of the anchor items for the TR table in the
anchor item design; NOT to be confused with the argument anchors that
is a vector of the anchor items CURRENT name for the analysis; if anchors
is not NULL, anchor_item_names may also only contain the same item name
mapping as anchors}
}
\description{
Summarize link DIF results
}
