% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{summary_dif}
\alias{summary_dif}
\title{Summary for DIF analysis}
\usage{
summary_dif(diflist, print = TRUE, save_at = NULL)
}
\arguments{
\item{diflist}{list with main and dif model as returned by dif_analysis()}

\item{print}{logical indicating whether summary is printed to the console}

\item{save_at}{character string; indicates the folder location where the
summaries are stored on the hard drive. Please note that the
path is relative to the current working path set by here::i_am(). Defaults
to NULL (not stored)}
}
\value{
list of information criteria, dif estimates and main effects in
  data frames for dif analysis
}
\description{
Summary for DIF analysis
}
