% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor_analysis.R
\name{summary_distractor_analysis}
\alias{summary_distractor_analysis}
\title{Summary of distractor analysis}
\usage{
summary_distractor_analysis(
  distractors,
  print = TRUE,
  save_at = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{distractors}{return object of distractor_analysis() function (list of
data frames containing item-total correlations for each item)}

\item{print}{logical indicating whether summary is printed to the console;
prints minimum and maximum for distractors and correct responses as well as
the respective item names}

\item{save_at}{character string; indicates the folder location where the
summaries are stored on the hard drive (first, the distractor object is
saved unchanged as Rdata; second, the distractor object is saved unchanged
as an excel file with one sheet per item distractor analysis; third, one
table for the distractors and one table for the correct responses is
saved as an excel file with one sheet per table). Please note that the
path is relative to the current working path set by here::i_am(). Defaults
to NULL (not stored)}

\item{overwrite}{boolean; indicates whether to overwrite existing file when saving table.}
}
\value{
list of data frames
          correct : item-total correlations for correct responses
          distractor : item-total correlations for distractors
}
\description{
Summary of distractor analysis
}
