% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{Tbl}
\alias{Tbl}
\title{Generate tables in NEPS Survey Papers}
\usage{
Tbl(
  obj,
  footnote = NULL,
  autofit = TRUE,
  merge = TRUE,
  lbl = NULL,
  hline = NULL,
  hline_head = NULL,
  size = 12,
  size_head = NULL,
  size_foot = 10,
  width = NULL,
  digits = 2,
  align = "center",
  align_head = "center"
)
}
\arguments{
\item{obj}{A data frame with the table content.}

\item{footnote}{A table note.}

\item{autofit}{A logical to automatically determine columns widths (\code{TRUE});
will be set to \code{FALSE}, if \code{width} is given.}

\item{merge}{A logical to merge header cells with the same content.}

\item{lbl}{A vector of labels for the header; the length of the vector
must correspond to the number of columns in \code{obj}.}

\item{hline}{A (named) vector of line numbers in the body for which a bottom
border is added; names of each element can be used to set the border color.}

\item{hline_head}{A (named) vector of line numbers in the header for which a
bottom border is added; names of each element can be used to set the border
color.}

\item{size}{The general font size.}

\item{size_head}{The font size of the header.}

\item{size_foot}{The font size of the footer.}

\item{width}{The widths of the columns; if a single value is provided, it
will be applied to all columns; otherwise the length must correspond to the
number of columns in \code{obj}.}

\item{digits}{A number for rounding; if a single number is provided, it will
be applied to all columns; otherwise the length must correspond to the
number of columns in \code{obj}.}

\item{align}{The alignment for each column in the body; if a single value is
provide, it will be applied to all columns; otherwise the length must
correspond to the number of columns in \code{obj}.}

\item{align_head}{the alignment for each column in the header; if a single
value is provided, it will be applied to all columns; otherwise the length
must correspond to the number of columns in \code{obj}.}
}
\value{
A flextable.
}
\description{
Create table
}
