% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblPars}
\alias{TblPars}
\title{Create table with item parameters}
\usage{
TblPars(
  obj,
  footnote = NULL,
  excl = c("N_administered"),
  size = 10,
  width = 0.5,
  rename_collapsed = TRUE
)
}
\arguments{
\item{obj}{A list with data frames with sheets from "irt_poly.xlsx"
or a data frame with sheet "summary" from "irt_poly.xlsx" created by
[irt_analysis()].}

\item{footnote}{A table note.}

\item{excl}{A vector of column names to exclude.}

\item{size}{The general font size.}

\item{width}{The widths of the columns; if a single value is provided, it
will be applied to all columns; otherwise the length must correspond to the
number of columns in `obj`.}

\item{rename_collapsed}{A boolean to remove the "_collapsed" suffix from
item names.}
}
\value{
A flextable.
}
\description{
Create table with item parameters
}
\examples{
\dontrun{
data(ex1)

# IRT analyses
tmpdir <- tempdir()
irt_analysis(
 resp = ex1$resp,
 vars = ex1$vars,
 select = "dich",
 valid = "valid",
 plots = FALSE,
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 warn = FALSE,
 verbose = FALSE,
)

# Import results of IRT analyses
pars <- Import(tmpdir, "/irt_dich.xlsx")

# Default parameter table
TblPars(pars)

# Parameter table including number of respondents the item was administered
# and a footnote
TblPars(pars, excl = NULL, footnote = "Nothing to report.")
}
}
