% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_setup.r
\name{Update}
\alias{Update}
\title{Update the Quarto extension in a folder}
\usage{
Update(path = getwd())
}
\arguments{
\item{path}{The path to the folder for the technical report.}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
Update the Quarto extension in a folder
}
\seealso{
\code{vignette("technical_report", package = "NEPSroutines")}
}
