% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{wright_map}
\alias{wright_map}
\title{Wright maps}
\usage{
wright_map(
  model,
  path = "Plots",
  name_group = NULL,
  width = 800,
  height = 1300
)
}
\arguments{
\item{model}{list; return object of irt_model()}

\item{path}{string; defines path to folder where plots shall be saved}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}

\item{width}{number; width of generated plot}

\item{height}{number; height of generated plot}
}
\description{
Create Wright maps for IRT models.
}
