% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{write_mvp_table}
\alias{write_mvp_table}
\title{Convert results of missing values by person to table}
\usage{
write_mvp_table(mv_p)
}
\arguments{
\item{mv_p}{list; return object of mv_person()}
}
\value{
table with frequency of missing values for one missing value type.
}
\description{
Convert results of missing values by person to table
}
