% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{write_results}
\alias{write_results}
\title{Convert results of missing values by person to table}
\usage{
write_results(res)
}
\arguments{
\item{res}{list with results as returned by mv_person}
}
\value{
table with frequency of missing values for one missing value type
}
\description{
Convert results of missing values by person to table
}
