% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age from birth and test date}
\usage{
calculate_age(
  resp,
  birth_year = "birthy",
  birth_month = "birthm",
  test_year = "testy",
  test_month = "testm",
  birth_day = NULL,
  test_day = NULL
)
}
\arguments{
\item{resp}{data.frame with birth and test date variables}

\item{birth_year}{string; contains name of variable with year of birth}

\item{birth_month}{string; contains name of variable with month of birth}

\item{test_year}{string or integer; contains either the name of the variable
in resp that includes the test year for each participant or one number with
the test year if it's the same for all participants}

\item{test_month}{string or integer; contains either the name of the variable
in resp that includes the test month for each participant or one number with
the test month if it's the same for all participants}

\item{birth_day}{string; contains name of variable with day of birth
(default is median)}

\item{test_day}{string or integer; contains either the name of the variable
in resp that includes the test day for each participant or one number with
the test day if it's the same for all participants (default is median)}
}
\value{
numeric vector with approximate age in years
}
\description{
Calculate age from birth and test date
}
\section{Notifications}{
\describe{
  \item{\code{message()} -- missing birth year values replaced by median}{
    Triggered when one or more \code{NA} values are present in the birth year
    variable. The affected values are replaced by the rounded sample median.
    Age calculations for those respondents are approximate.
  }
  \item{\code{message()} -- missing birth month values replaced by median}{
    Triggered when one or more \code{NA} values are present in the birth month
    variable. The affected values are replaced by the rounded sample median.
    Age calculations for those respondents are approximate.
  }
  \item{\code{message()} -- missing test year values replaced by median}{
    Triggered when one or more \code{NA} values are present in the test year
    variable. The affected values are replaced by the rounded sample median.
    Age calculations for those respondents are approximate.
  }
  \item{\code{message()} -- missing test month values replaced by median}{
    Triggered when one or more \code{NA} values are present in the test month
    variable. The affected values are replaced by the rounded sample median.
    Age calculations for those respondents are approximate.
  }
}
}
