% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age from birth and test date}
\usage{
calculate_age(
  resp,
  birth_year,
  birth_month,
  birth_day = NULL,
  test_year,
  test_month,
  test_day = NULL
)
}
\arguments{
\item{resp}{data.frame with birth and test date variables}

\item{birth_year}{string; contains name of variable with year of birth}

\item{birth_month}{string; contains name of variable with month of birth}

\item{birth_day}{string; contains name of variable with day of birth (default is 15)}

\item{test_year}{string; contains name of variable with year of test}

\item{test_day}{string; contains name of variable with day of test (default is 15)}

\item{tets_month}{string; contains name of variable with month of test}
}
\value{
numeric vector with approximate age in years
}
\description{
Calculate age from birth and test date
}
