% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{calculate_link_parameters}
\alias{calculate_link_parameters}
\title{calculate link constant and link error for anchor group or items design}
\usage{
calculate_link_parameters(
  is_anchor_items = TRUE,
  select_previous,
  select_current,
  anchors = NULL,
  xsi_previous,
  xsi_current,
  select_link_sample = NULL,
  xsi_link_sample = NULL,
  vars,
  scoring = "scoring"
)
}
\arguments{
\item{is_anchor_items}{logical; whether anchor items (TRUE) or anchor group
(FALSE) design is used}

\item{select_previous}{character; contains name of logical variable in vars
identifying the item set for the previous measurement time point}

\item{select_current}{character; contains name of logical variable in vars
identifying the item set for the current measurement time point}

\item{anchors}{character vector of selected anchor items}

\item{xsi_previous}{matrix; item difficulty parameters of previous
measurement}

\item{xsi_current}{matrix; item difficulty parameters of current measurement}

\item{select_link_sample}{string; defines name of logical variable in vars
that indicates the items for the link sample analysis}

\item{xsi_link_sample}{matrix; item difficulty parameters of link sample}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to the link constant (e.g., only
half the link constant is needed if a PC item is scored 0.5)}
}
\value{
list(const, const.err, anchors_previous, anchors_current,
  xsi_current.linked)
}
\description{
calculate link constant and link error for anchor group or items design
}
