% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{check_dif_anchor}
\alias{check_dif_anchor}
\title{Test measurement invariance / DIF over time}
\usage{
check_dif_anchor(
  resp_previous,
  resp_current,
  resp_link_sample = NULL,
  anchors = NULL,
  mvs = NULL,
  return = TRUE,
  valid,
  scoring = "scoring",
  vars,
  items = NULL,
  select_previous,
  select_current,
  diff_threshold = 0.5,
  anchor_item_names = NULL
)
}
\arguments{
\item{resp_previous}{data.frame with responses of first measurement wave;
contains ID_t; anchor items are named to fit the current analysis}

\item{resp_current}{data.frame with responses of second measurement wave;
contains ID_t}

\item{resp_link_sample}{data.frame with responses of link sample;
contains ID_t; required if anchor group design is used}

\item{anchors}{data.frame with two columns including the link items
in the two waves; if NULL, all common items are used}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{return}{logical; whether results shall be returned}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to the link constant (e.g., only
half the link constant is needed if a PC item is scored 0.5)}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{items}{character; contains name of logical variable in vars
identifying the item set for the both time points}

\item{select_previous}{character; contains name of logical variable in vars
identifying the item set for the previous measurement time point; needed to
build tables}

\item{select_current}{character; contains name of logical variable in vars
identifying the item set for the current measurement time point; needed to
build tables}

\item{diff_threshold}{numeric; threshold under which DIF in common link items
is accepted; defaults to .5}

\item{anchor_item_names}{character matrix; contains the old (column 1) and
new (column 2) item names of the anchor items for the TR table in the
anchor item design; NOT to be confused with the argument anchors that
is a vector of the anchor items CURRENT name for the analysis; if anchors
is not NULL, anchor_item_names may also only contain the same item name
mapping as anchors}
}
\value{
list with DIF model(s), DIF summary and anchor items
}
\description{
Test measurement invariance / DIF over time
}
