% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_folder}
\alias{check_folder}
\title{Check if folder exists and if not, create new one}
\usage{
check_folder(path)
}
\arguments{
\item{path}{path to folder}
}
\description{
Check if folder exists and if not, create new one
}
