% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_logicals}
\alias{check_logicals}
\title{Check logical variable(s) for correctness}
\usage{
check_logicals(df, name_df, logicals, warn = TRUE)
}
\arguments{
\item{df}{data.frame; contains at least the indicated logical variable(s)}

\item{name_df}{string; contains name of data.frame}

\item{logicals}{character vector; contains names of variables that shall be
checked for correctness}

\item{warn}{logical; whether to warn if logicals include NA}
}
\description{
Check logical variable(s) for correctness
}
