% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{collapse_response_categories}
\alias{collapse_response_categories}
\title{Collapse response categories with N < 200}
\usage{
collapse_response_categories(
  resp,
  vars,
  select_poly,
  per_cat = 200,
  path_table = here::here("Tables"),
  print = TRUE,
  save = FALSE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select_poly}{character; indicates the logical variable in vars which
contains the item names of the polytomous items}

\item{per_cat}{integer; minimum number of persons per category; defaults to 200}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{print}{logical; whether results shall be printed to console}

\item{save}{logical; whether results shall be saved to hard drive}
}
\value{
resp with collapsed categories. Note that this function changes the
  given PC items IN PLACE. If you want to keep the original data, please
  copy and rename the items to be collapsed first.
}
\description{
Collapse response categories with N < 200
}
