% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{conduct_dif_analysis}
\alias{conduct_dif_analysis}
\title{Conduct DIF analyses}
\usage{
conduct_dif_analysis(items, dif_vars, resp, vars, scoring, valid, mvs, verbose)
}
\arguments{
\item{items}{string; defines name of logical variable in vars that indicates
which items to use for the analysis; if some of the \code{dif_vars}
come with a different set of analysis items, this argument becomes a
vector of \code{length(dif_vars)} containing the respective selection
variables in vars}

\item{dif_vars}{character vector; contains the variable names to be tested
for DIF (e.g., "gender")}

\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; defaults to
"scoring"}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{verbose}{logical; whether to print processing information to console}
}
\value{
a list of length(dif_vars) lists containing each:
  mmod: main effects model
  dmod: DIF effects model
}
\description{
Testing for differential item functioning for binary and polytomous data.
Main effects and DIF effects models are estimated.
}
