% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{conduct_dif_analysis}
\alias{conduct_dif_analysis}
\title{Conduct DIF analyses}
\usage{
conduct_dif_analysis(
  resp,
  vars,
  select,
  dif_vars,
  valid = NULL,
  scoring = NULL,
  mvs = NULL,
  include_mv = 200,
  path = here::here("Results"),
  save = TRUE,
  name_group = NULL,
  control = NULL,
  pweights = NULL,
  verbose = FALSE,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{character vector; defines name(s) of logical variable(s) in vars
that indicates which items to use for analysis; if some of the \code{dif_vars}
come with a different set of analysis items, this argument becomes a
vector of \code{length(dif_vars)} containing the respective selection
variables in vars}

\item{dif_vars}{character vector; contains the variable names to be tested
for DIF (e.g., "gender")}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{include_mv}{numeric; identifies threshold for which group size missing
values should be included in analysis as an extra group (defaults to 200)}

\item{path}{string; defines path to folder where results shall be saved}

\item{save}{logical; whether results shall be saved to hard drive}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}

\item{control}{list; function argument as passed to TAM-functions}

\item{pweights}{character; defines name of numerical variable in resp that
contains person weights as passed to TAM-function}

\item{verbose}{logical; whether to print processing information to console}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\value{
a list of length(dif_vars) lists containing each:
  mmod: main effects model
  dmod: DIF effects model
}
\description{
Testing for differential item functioning for binary and polytomous data.
Main effects and DIF effects models are estimated.
}
