% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_analysis.R
\name{conduct_dim_analysis}
\alias{conduct_dim_analysis}
\title{Conduct dimensionality analysis}
\usage{
conduct_dim_analysis(
  resp,
  vars,
  items,
  scoring,
  dim,
  valid,
  irtmodel,
  maxiter,
  snodes,
  verbose,
  mvs
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{items}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; defaults to
"scoring"}

\item{dim}{character vector. contains names of all dimension variables
that shall be analyzed (NOTE: the dimensions must be coded as integers from 1
to the number of dimensions!)}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{irtmodel}{string; "1PL" for Rasch, "2PL" for 2PL, "PCM2" for PCM and
"GPCM" for GPCM analysis}

\item{maxiter}{max iterations as passed to the TAM function}

\item{snodes}{snodes as passed to the TAM function}

\item{verbose}{verbose as passed to the TAM function}

\item{mvs}{named integer vector; contains user-defined missing values}
}
\value{
list of results for each dimensional analysis
}
\description{
Conduct dimensionality analysis
}
