% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor_analysis.R
\name{conduct_dis_analysis}
\alias{conduct_dis_analysis}
\title{Conduct distractor analysis}
\usage{
conduct_dis_analysis(
  resp,
  vars,
  valid = NULL,
  select_raw,
  select_score = "dich",
  correct = "correct_response",
  mvs = NULL,
  warn = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variables 'item' (character) containing item names,
                   'dich' (logical) indicating all dichotomously scored items
                   'raw' (logical) indicating all unscored items,
                   'type' (character) containing item type (e.g. MC),
                   'correct_response' (integer) containing the correct
                     response for each item,
and all variables that are further defined in the function arguments}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{select_raw}{string; defines name of logical variable in vars that
indicates all raw items that shall be used for the distractor analysis}

\item{select_score}{string; defines name of logical variable in vars that
indicates all items that shall be used to caculate total score}

\item{correct}{string; defines name of variable in vars that contains the
correct responses to the items}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}
}
\description{
Calculates the item-rest correlations of distractor and correct responses
}
