% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor_analysis.R
\name{conduct_dis_analysis}
\alias{conduct_dis_analysis}
\title{Conduct distractor analysis}
\usage{
conduct_dis_analysis(resp, vars, items, valid = NULL)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variables 'items' (character) containing item names,
                   'raw' (logical) indicating unscored items,
                   'type' (character) containing item type (e.g. MC),
                   'correct_response' (integer) containing the correct
                     response for each item,
and all variables that are further defined in the function arguments}

\item{items}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}
}
\description{
Calculates the item-rest correlations of distractor and correct responses
}
