% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_mv}
\alias{convert_mv}
\title{Convert user-defined missing values to NAs}
\usage{
convert_mv(resp, items = NULL, mvs = NULL, warn = TRUE)
}
\arguments{
\item{resp}{data.frame containing the item responses}

\item{items}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{warn}{logical; whether warnings are to be printed to the console}
}
\value{
data.frame like resp, but without user-defined mvs
}
\description{
User defined missing values (usually coded as negative numbers) must be
converted to NAs so that they do not break the IRT analysis
}
