% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scores.R
\name{create_scores}
\alias{create_scores}
\title{Create scores}
\usage{
create_scores(
  resp,
  vars,
  scoring = NULL,
  score_name = "score",
  xsi_fixed = NULL,
  facet = NULL,
  select,
  valid = NULL,
  mvs = NULL,
  wle = TRUE,
  sum_score = FALSE,
  control_tam = NULL,
  control_wle = NULL,
  pweights = NULL,
  poly2dich = TRUE,
  return = FALSE,
  do_dim = TRUE,
  do_dif = TRUE,
  diff_threshold = 0.5,
  wid = NULL,
  snodes = 0,
  maxiter = 1000,
  digits = 3,
  verbose = TRUE,
  warn = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; can be NULL for
Rasch model}

\item{score_name}{character; name of the scores -- WITHOUT extension (e.g.,
reg4 instead of reg4_sc1 or mag12 instead of mag12_sc1u)}

\item{xsi_fixed}{matrix; fixed parameter estimates of final IRT scaling as
returned by tam.mml()}

\item{facet}{character vector; contains the variable name indicating the
test rotation}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{wle}{logical; whether to estimate WLEs}

\item{sum_score}{logical; whether to calculate sum scores}

\item{control_tam}{list; control argument as passed to tam.mml.mfr()}

\item{control_wle}{list; can contain Msteps and/or convM as to pass to tam.wle()
as elements of the list}

\item{pweights}{numeric vector; person weights for current measurement point
passed to TAM-functions}

\item{poly2dich}{logical; indicates whether count only correctly scored
binary or FULLY correctly scored PC items}

\item{return}{logical; whether results shall be returned}

\item{do_dim}{logical; whether to do dimensionality analysis for linking}

\item{do_dif}{logical; whether to do dif analysis for linking}

\item{diff_threshold}{numeric; threshold under which DIF in common link items
is accepted; defaults to .5}

\item{wid}{variable name used as WLE identifier in first measurement wave}

\item{snodes}{snodes as passed to the TAM function for the dimensionality
analyses}

\item{maxiter}{maximum number of iterations as passed to the TAM function
for the dimensionality analyses}

\item{digits}{numeric; number of decimal places for rounding}

\item{verbose}{logical; verbose as passed to the TAM function}

\item{warn}{logical; whether to print warnings}

\item{resp_previous}{data.frame with responses of first measurement wave,
a person identifier; may also include a WLE as given in wid}

\item{resp_link_study}{data.frame with responses of link sample,
a person identifier;}

\item{pweights_previous}{numeric vector; person weights for previous
measurement point passed to TAM-functions}

\item{pweights_link_study}{numeric vector; person weights for link study
passed to TAM-functions}

\item{scoring_previous}{numeric; named vector with the scoring factors to
be applied to the loading matrix of the previous measurement point; can be
NULL for the Rasch model}

\item{scoring_link_study}{numeric; named vector with the scoring factors to
be applied to the loading matrix of the link study; can be NULL for the
Rasch model}

\item{anchors}{character; data.frame with two columns including the link
items; for anchor item designs the first column refers to the previous
measurement time point and the second column to the current measurement
time point; for anchor group designs the first column refers to the main
sample (for both time points) and the second column refers to the link
sample; if NULL, all common items are used}

\item{longitudinal}{logical; do within cohort linking (TRUE) or between
cohort linking (FALSE)}

\item{print}{logical; whether results shall be printed to console}

\item{save}{logical; whether results shall be saved to hard drive}

\item{overwrite}{logical; whether to overwrite existing file when saving
table}

\item{path_results}{string; defines path to folder where results shall be
saved}

\item{path_table}{string; defines path to folder where tables shall be saved}
}
\description{
Create scores
}
