% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_suf_and_sum_scores.R
\name{create_suf}
\alias{create_suf}
\title{Create SUF}
\usage{
create_suf(
  resp,
  vars,
  items_suf,
  wle_name,
  xsi_fixed = NULL,
  rotated = FALSE,
  linked = FALSE,
  wles = NULL,
  linked_wles = NULL,
  linked_location = NULL,
  facet = NULL,
  items = NULL,
  valid = NULL,
  mvs = NULL,
  scoring = NULL,
  competence,
  filename = "suf",
  path_results = here::here("Results"),
  save = TRUE,
  return = FALSE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{items_suf}{character; indicates the logical variable in vars which
contains the item names designated for the SUF (original scored items
before the first IRT analyses)}

\item{wle_name}{character; name of the wle -- WITHOUT extension (e.g.,
reg4 instead of reg4_sc1 or mag12 instead of mag12_sc1u)}

\item{xsi_fixed}{matrix; fixed parameter estimates of final IRT scaling as
returned by tam.mml()}

\item{rotated}{logical; whether the test design contains rotated tests; if
TRUE, the WLEs have to be estimated again}

\item{linked}{logical; whether the test is linked to a previous assessment}

\item{wles}{data.frame; contains ID_t, WLEs and SEs as estimated by tam.wle()
for the current assessment}

\item{linked_wles}{data.frame; contains the linked WLEs as returned by
link_samples() / linking()}

\item{linked_location}{character; path to linking results as saved by
linking(); alternative to linked_wles}

\item{facet}{character vector; contains the variable name indicating the
test rotation}

\item{items}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; can be NULL for
Rasch model}

\item{competence}{character; name of tested competence for labelling the
data (capitalized!)}

\item{filename}{character; indicates final name of SUF according to study}

\item{path_results}{character; indicates where SUF shall be saved}

\item{save}{logical; whether results shall be saved to hard drive}

\item{return}{logical; whether results shall be returned}
}
\description{
Create SUF
}
