% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_suf.R
\name{create_suf}
\alias{create_suf}
\title{Create SUF}
\usage{
create_suf(
  resp,
  vars,
  select,
  competence,
  scores = NULL,
  mvs = NULL,
  score_name = NULL,
  items_labels = NULL,
  suf_name = "suf",
  name_group = NULL,
  path = "suf",
  save = TRUE,
  return = TRUE,
  warn = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{character; indicates the logical variable in vars which
contains the item names designated for the SUF (original scored items
before the first IRT analyses)}

\item{competence}{character; name of tested competence for labeling the
data}

\item{scores}{data.frame; contains test scores (potentially with standard
errors) as variables and persons as rows; additionally includes ID_t as a
person identifier; return object of functions c}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{score_name}{character; name of the scores -- WITHOUT extension (e.g.,
reg4 instead of reg4_sc1 or mag12 instead of mag12_sc1u)}

\item{items_labels}{vector; named vector for each item used for labeling
the data}

\item{suf_name}{character; name of suf (defaults to 'suf')}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}

\item{path}{character; indicates where SUF shall be saved}

\item{save}{logical; whether results shall be saved to hard drive}

\item{return}{logical; whether results shall be returned}

\item{warn}{logical; whether to print warnings}
}
\description{
Create SUF
}
