% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desc_abs}
\alias{desc_abs}
\title{Show frequency (absolute) of answers of list of variables}
\usage{
desc_abs(resp, desc, valid = NULL, warn = TRUE)
}
\arguments{
\item{resp}{data.frame; contains sociodemographic and booklet variables}

\item{desc}{character vector; contains the name(s) of the variable(s) that
shall be used in analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{warn}{logical; whether to warn if no parameter 'valid' provided}
}
\value{
list (table) with  absolute frequencies for each variable in 'desc'.
}
\description{
Show frequency (absolute) of answers of list of variables
}
