% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desc_abs}
\alias{desc_abs}
\title{Show frequency (absolute) of answers of list of variables}
\usage{
desc_abs(dat, desc, valid = NULL)
}
\arguments{
\item{dat}{data.frame with item responses (user-defined missing values)
and logical grouping variables, named after the grouping
factor (e.g. "easy" and "difficult" in the case of testlet.}

\item{desc}{character. contains names of variables for descriptive analysis.}

\item{valid}{character string. defines name of boolean variable in dat,
indicating (in)valid cases.}
}
\value{
list with  frequency (absolute) of answers for each variable in desc.
}
\description{
Show frequency (absolute) of answers of list of variables
}
