% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desc_con}
\alias{desc_con}
\title{Descriptives of continuous variables}
\usage{
desc_con(resp, desc, valid = NULL, digits = 3, print = TRUE, return = FALSE)
}
\arguments{
\item{resp}{data.frame; contains sociodemographic and booklet variables}

\item{desc}{character vector; contains the name(s) of the variable(s) that
shall be used in analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{digits}{integer; number of decimals for rounding}

\item{print}{logical; whether results shall be printed to console}

\item{return}{logical; whether results shall be returned}
}
\value{
list with descriptive statistics (n, mean, etc.) for each variable
defined in 'desc'.
}
\description{
Descriptives of continuous variables
}
