% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desc_nom}
\alias{desc_nom}
\title{Descriptives of nominal variables}
\usage{
desc_nom(
  dat,
  desc,
  min.val = 3,
  digits = 1,
  attributes = TRUE,
  absolute = TRUE,
  percentage = TRUE
)
}
\arguments{
\item{dat}{data.frame with item responses (user-defined missing values)
and logical grouping variables, named after the grouping
factor (e.g. "easy" and "difficult" in the case of testlet.}

\item{desc}{character. contains names of variables for descriptive analysis.}

\item{min.val}{minimum number of valid values;
if negative, set to the default of 3}

\item{digits}{number of decimals for rounding percentages}

\item{attributes}{???}

\item{absolute}{???}

\item{percentage}{???}
}
\value{
list with descriptives.
}
\description{
Descriptives of nominal variables
}
