% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desc_perc}
\alias{desc_perc}
\title{Show frequency (relative) of answers of list of variables}
\usage{
desc_perc(dat, desc, valid = NULL, digits = 1)
}
\arguments{
\item{dat}{data.frame with item responses (user-defined missing values)
and logical grouping variables, named after the grouping
factor (e.g. "easy" and "difficult" in the case of testlet.}

\item{desc}{character. contains names of variables for descriptive analysis.}

\item{valid}{character string. defines name of boolean variable in dat,
indicating (in)valid cases.}

\item{digits}{number of decimals for rounding}
}
\value{
list with frequency (relative) of answers for each variable in desc.
}
\description{
Show frequency (relative) of answers of list of variables
}
