% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desc_perc}
\alias{desc_perc}
\title{Show frequency (relative) of answers of list of variables}
\usage{
desc_perc(dat, desc, valid = NULL, warn = TRUE, useNA = "always", digits = 1)
}
\arguments{
\item{dat}{data.frame; contains sociodemographic and booklet variables}

\item{desc}{character vector; contains the name(s) of the variable(s) that
shall be used in analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{warn}{logical; whether to warn if no parameter 'valid' provided}

\item{useNA}{string; "no" if NAs should not be included in table, "always"
(default) if NAs should be included in table}

\item{digits}{integer; number of decimals for rounding}
}
\value{
list with frequency (relative) of answers for each variable in desc.
}
\description{
Show frequency (relative) of answers of list of variables
}
