% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{dichotomous_scoring}
\alias{dichotomous_scoring}
\title{Dichotomous scoring of MC items}
\usage{
dichotomous_scoring(resp, vars, select, correct_response)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which (unscored) items to use for the analysis}

\item{correct_response}{character; indicates the logical variable in vars
which contains the correct response of the MC items}
}
\value{
resp with dichotomously scored MC items. Note that this function
  changes the given MC items IN PLACE. If you want to keep the original data,
  please copy and rename the items to be scored first.
}
\description{
Dichotomous scoring of MC items
}
