% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{dichotomous_scoring}
\alias{dichotomous_scoring}
\title{Dichotomous scoring of MC items}
\usage{
dichotomous_scoring(resp, vars, old_names, new_names = NULL, sep = ";")
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{old_names}{character vector; contains the names of the original items}

\item{new_names}{character vector; contains the names of the new items. Must
be in same order as parameter "old_names"! (default is old name + "_c")}

\item{sep}{string; defines type of punctuation used for separating several
correct responses in the variable correct_response in vars (default is ";")}
}
\value{
resp with dichotomously scored MC items.
}
\description{
Dichotomous scoring of MC items
}
