% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{dif_all}
\alias{dif_all}
\title{DIF analyses - all in one function}
\usage{
dif_all(
  resp,
  vars,
  items,
  dif_vars,
  scoring = NULL,
  path_table = "Tables",
  print_table = FALSE,
  return_results = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{resp}{data.frame containing the responses. y in {0, 1} for binary data
and y in {0, 1, ... k-1} for polytomous responses with k categories.
Missing values (-97 to -21) are coded as NA internally. Also contains a
column with the DIF variable.}

\item{vars}{data.frame with all variables as rows with at least following columns:
items: contains all item names (both scored and unscored)
scored: logical or binary integer, identifies all scored items}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{dif_vars}{character vector. contains the variable names to be tested for DIF
(e.g., "gender")}

\item{scoring}{numeric vector; scoring factor to be applied to loading matrix;
can be NULL for Rasch model}

\item{path_table}{???}

\item{print_table}{???}

\item{return_results}{boolean. indicates whether to return results.}

\item{verbose}{logical; should progress be printed to console?}

\item{...}{additional arguments to be passed to tam.mml}
}
\value{
a list of:
  mmod: main effects model
  dmod: DIF effects model
}
\description{
Testing for differential item functioning for binary and polytomous data.
Main effects and DIF effects models are estimated.
}
