% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{dif_model}
\alias{dif_model}
\title{DIF analyses}
\usage{
dif_model(
  resp,
  vars,
  items,
  facets,
  scoring = "scoring",
  valid = NULL,
  verbose = FALSE
)
}
\arguments{
\item{resp}{data.frame containing the responses. y in {0, 1} for binary data
and y in {0, 1, ... k-1} for polytomous responses with k categories.
Missing values (-97 to -21) are coded as NA internally. Also contains a
column with the DIF variable.}

\item{vars}{data.frame with all variables as rows with at least following columns:
items: contains all item names (both scored and unscored)
further custom-named columns indicating the items to select and the item
scoring via logicals}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{facets}{character string signifying the variable to be tested for DIF
(e.g., "gender")}

\item{scoring}{numeric vector; scoring factor to be applied to loading matrix;
can be NULL for Rasch model}

\item{valid}{character string. defines name of boolean variable in data,
indicating (in)valid cases.}

\item{verbose}{logical; should progress be printed to console?}
}
\value{
a list of:
  mmod: main effects model
  dmod: DIF effects model
}
\description{
Testing for differential item functioning for binary and polytomous data.
Main effects and DIF effects models are estimated.
}
