% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{dif_model}
\alias{dif_model}
\title{DIF analyses}
\usage{
dif_model(
  resp,
  vars,
  select,
  dif_var,
  scoring = NULL,
  valid = NULL,
  include_mv = 200,
  mvs = NULL,
  verbose = FALSE,
  warn = TRUE,
  test = TRUE,
  control = NULL,
  pweights = NULL
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis; if some of the \code{dif_vars}
come with a different set of analysis items, this argument becomes a
vector of \code{length(dif_vars)} containing the respective selection
variables in vars}

\item{dif_var}{string; defines the name of the variable to be tested for DIF
(e.g., "gender")}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{include_mv}{numeric; identifies threshold for which group size missing
values should be included in analysis as an extra group (defaults to 200)}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{verbose}{logical; whether to print processing information to console}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}

\item{test}{logical; whether to test data structure (should be set to TRUE)}

\item{control}{list; function argument as passed to TAM-functions}

\item{pweights}{character; defines name of numerical variable in resp that
contains person weights as passed to TAM-function}

\item{two_par}{logical; whether two parameter model (2PL or GPCM) shall be
used as base for DIF analyses (defaults to FALSE)}
}
\value{
a list of:
  mmod: main effects model
  dmod: DIF effects model
}
\description{
Testing for differential item functioning for binary and polytomous data.
Main effects and DIF effects models are estimated.
}
