% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{dif_summary}
\alias{dif_summary}
\title{Summary for DIF analysis}
\usage{
dif_summary(diflist, print = TRUE, save_at = NULL, overwrite = FALSE)
}
\arguments{
\item{diflist}{list with main and dif model as returned by dif_model()}

\item{print}{logical indicating whether summary is printed to the console}

\item{save_at}{character string; indicates the folder location where the
summaries are stored on the hard drive. Please note that the
path is relative to the current working path set by here::i_am(). Defaults
to NULL (not stored)}

\item{overwrite}{boolean; indicates whether to overwrite existing file when saving table.}
}
\value{
list of information criteria, dif estimates and main effects in
  data frames for dif analysis
}
\description{
Summary for DIF analysis
}
