% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{dif_summary}
\alias{dif_summary}
\title{Summary for DIF analysis}
\usage{
dif_summary(
  diflist,
  print = TRUE,
  save = TRUE,
  path = here::here("Tables"),
  prob_dif = 0.5,
  overwrite = FALSE
)
}
\arguments{
\item{diflist}{list; return object of dif_model(); with main and dif model}

\item{print}{logical; whether results shall be printed to console}

\item{save}{logical; whether results shall be saved to hard drive}

\item{path}{string; defines path to folder where tables shall be saved}

\item{prob_dif}{numeric scalar; indicates absolute threshold of problematic DIF (defaults to 0.5)}

\item{overwrite}{logical; whether to overwrite existing file when saving table}
}
\value{
list of information criteria, dif estimates and main effects in
  data frames for dif analysis
}
\description{
Summary for DIF analysis
}
