% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_analysis.R
\name{dim_summary}
\alias{dim_summary}
\title{Summary of dimensionality analysis (saves summary in excel sheet)}
\usage{
dim_summary(dimensionality)
}
\arguments{
\item{dimensionality}{list with results of the dim_analysis function}
}
\value{
data.frame of summary results
}
\description{
Summary of dimensionality analysis (saves summary in excel sheet)
}
