% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor_analysis.R
\name{dis_analysis}
\alias{dis_analysis}
\title{Distractor analysis}
\usage{
dis_analysis(
  resp,
  vars,
  valid = NULL,
  mvs = NULL,
  select_raw,
  select_score = "dich",
  use_wle = FALSE,
  scoring = NULL,
  correct = "correct_response",
  name_group = NULL,
  save = TRUE,
  print = TRUE,
  return = FALSE,
  path_results = "Results",
  path_table = "Tables",
  overwrite = FALSE,
  digits = 3,
  warn = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in \{0, 1\} for binary data and y in \{0, 1, ... k-1\} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variables 'item' (character) containing item names and all variables
that are further defined in the function arguments}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{select_raw}{string; defines name of logical variable in vars that
indicates all raw items that shall be used for the distractor analysis}

\item{select_score}{string; defines name of logical variable in vars that
indicates all items that shall be used to calculate total score}

\item{use_wle}{logical; use WLE instead of sum score for item-total
correlations}

\item{scoring}{string; defines name of scoring variable in vars}

\item{correct}{string; defines name of variable in vars that contains the
correct responses to the items}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}

\item{save}{logical; whether results shall be saved to hard drive}

\item{print}{logical; whether results shall be printed to console}

\item{return}{logical; whether results shall be returned}

\item{path_results}{string; indicates the folder location where the summaries
are stored on the hard drive}

\item{path_table}{string; indicates the folder location where the tables
are stored on the hard drive}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{digits}{integer; number of decimals for rounding}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}
}
\value{
(if return = TRUE) list with results:
analysis: list with one data frame per item containing item-total correlations
for each possible response; correct response is marked with an *
summary: list with summary of results.
}
\description{
Checks how the distractors in multiple choice items functioned.
}
