% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor_analysis.R
\name{dis_analysis}
\alias{dis_analysis}
\title{Distractor analysis}
\usage{
dis_analysis(
  resp,
  vars,
  valid = NULL,
  item_type = "MC",
  scored_version = NULL,
  mvs = NULL,
  save = TRUE,
  print = TRUE,
  return = FALSE,
  path_results = here::here("Results"),
  path_table = here::here("Tables"),
  overwrite = FALSE,
  warn = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variables 'item' (character) containing item names,
                   'raw' (logical) indicating unscored items,
                   'type' (character) containing item type (e.g. MC),
                   'correct_response' (integer) containing the correct
                     response for each item,
and all variables that are further defined in the function arguments}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{item_type}{character vector; contains all item types as specified in
vars$type that shall be used for distratcor analysis (default is 'MC')}

\item{scored_version}{string; defines name of variable in vars that contains
the names of the scored versions of the raw items; can be left empty, then
default of name_scored_item = name_unscored_item + "_c" applies}

\item{save}{logical; whether results shall be saved to hard drive}

\item{print}{logical; whether results shall be printed to console}

\item{return}{logical; whether results shall be returned}

\item{path_results}{string; indicates the folder location where the summaries
are stored on the hard drive; please note that the path is relative to the
current working path set by here::i_am()}

\item{path_table}{string; indicates the folder location where the tables
are stored on the hard drive; please note that the path is relative to the
current working path set by here::i_am()}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}
}
\value{
list of one data frame per item containing item-total correlations
  for each possible response; correct response is marked with an *
}
\description{
Checks how the distractors in multiple choice items functioned.
}
