% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor_analysis.R
\name{dis_summary}
\alias{dis_summary}
\title{Summary of distractor analysis}
\usage{
dis_summary(distractors)
}
\arguments{
\item{distractors}{return object of dis_analysis() function (list of
data frames containing item-total correlations for each item)}
}
\value{
list of data frames
          correct : item-total correlations for correct responses
          distractor : item-total correlations for distractors
}
\description{
Summary of distractor analysis
}
