% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{duplicate_items}
\alias{duplicate_items}
\title{Duplicate item information in vars}
\usage{
duplicate_items(vars, old_names, new_names, change = NULL)
}
\arguments{
\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{old_names}{character vector; contains the names of the original items}

\item{new_names}{character vector; contains the names of the new items. Must
be in same order as parameter "old_names"!}

\item{change}{named character vector; if some information in vars about the
new items shall be changed, include the respective variable as the name and the new
value as the value of the vector (e.g. change = c(raw = FALSE, dich = TRUE)),
if raw shall be set to FALSE and dich to TRUE for all new items.}
}
\value{
vars with new rows including the duplicated items.
}
\description{
Duplicate item information in vars
}
\section{Notifications}{

\describe{
\item{\code{stop}}{One or more names in \code{old_names} are not found in
\code{vars$item}. Effect: the function halts; no new rows are added to
\code{vars}.}
}
}

