% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_figures.r
\name{FigWrightMap}
\alias{FigWrightMap}
\alias{FigWright_Map}
\title{Crop plot with Wright map and include footnote}
\usage{
FigWrightMap(file, tbl = 5)

FigWright_Map(file, tbl = 5)
}
\arguments{
\item{file}{A file path for the image.}

\item{tbl}{The table number for the footnote.}
}
\value{
An image.
}
\description{
Crop plot with Wright map and include footnote
}
