% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetCat}
\alias{GetCat}
\title{Summarize category thresholds}
\usage{
GetCat(obj, stat = median, item = FALSE, digits = 2, rename_collapsed = TRUE)
}
\arguments{
\item{obj}{A [TAM::tam.mml()] object with a fitted IRT model from
"irt_dich.rds" or "irt_poly.rds" generated by[irt_analysis()].}

\item{stat}{A function used to summarize the category thresholds.}

\item{item}{A logical indicating whether to return the results from `stat`
(`FALSE`) or the item name corresponding to the value of `stat` (`TRUE`).}

\item{digits}{A number for rounding.}

\item{rename_collapsed}{A boolean to remove the "_collapsed" suffix from
item names.}
}
\value{
The calculated statistic or a vector of item names.
}
\description{
Summarize category thresholds
}
\examples{
\dontrun{
data("ex2")

# IRT analyses
tmpdir <- tempdir()
irt_analysis(
 resp = ex2$resp,
 vars = ex2$vars,
 select = "mixed",
 valid = "valid",
 plots = FALSE,
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 warn = FALSE,
 verbose = FALSE,
)

# Import results of IRT analyses
irt <- readRDS(paste0(tmpdir, "/irt_poly.rds"))

# Largest category threshold
GetCat(irt, max)

# Item with the largest category threshold
GetCat(irt, max, item = TRUE)

# Clean up generated files
file.remove(paste0(tmpdir, "/irt_poly.xlsx"))
file.remove(paste0(tmpdir, "/irt_poly.rds"))
}
}
