% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetCat}
\alias{GetCat}
\title{Summarize category thresholds}
\usage{
GetCat(obj, stat = median, item = FALSE, digits = 2)
}
\arguments{
\item{obj}{A [TAM::tam.mml()] object with a fitted IRT model from
"irt_dich.rds" or "irt_poly.rds" generated by[scaling::irt_analysis()].}

\item{stat}{A function used to summarize the category thresholds.}

\item{item}{A logical indicating whether to return the results from `stat`
(`FALSE`) or the item name corresponding to the value of `stat` (`TRUE`).}

\item{digits}{A number for rounding.}
}
\value{
The calculated statistic or a vector of item names.
}
\description{
Summarize category thresholds
}
