% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetDif}
\alias{GetDif}
\alias{GetDIF}
\title{Get values from DIF analyses}
\usage{
GetDif(
  obj,
  n = NULL,
  main = NULL,
  dif = NULL,
  item = FALSE,
  group = NULL,
  model = "dif",
  signed = FALSE,
  digits = 2
)

GetDIF(
  obj,
  n = NULL,
  main = NULL,
  dif = NULL,
  item = FALSE,
  group = NULL,
  model = "dif",
  signed = FALSE,
  digits = 2
)
}
\arguments{
\item{obj}{A data frame with sheets from "dif_dich_XXX.xlsx" or
"dif_poly_XXX.xlsx" created by [scaling::dif_analysis()].}

\item{n}{A group name to return the sample size of this group
(corresponds to the values in column "number" of sheet "facet").}

\item{main}{Return standardized main effects (`std`) or unstandardized
main effects (`ustd`). If the DIF variable has more than two groups
and `group` is missing, results for the first group comparison are returned.}

\item{dif}{A function used to summarize the DIF effects or a string
specifying an operator (=, <, >) with a number; e.g.,
* =1 returns number of DIF effects equal to 1
* <1 returns number of DIF effects less than 1
* >1 returns number of DIF effects greater than 1
Multiple conditions can be combined using & and |.
If the DIF variable has more than two group and `group` is missing, results
across all groups are returned.}

\item{item}{A logical to return the item names (`TRUE`) instead of the
calculated statistic for the DIF effects (`FALSE`).}

\item{group}{If the DIF variable has more than two groups, the name of
the group comparison.}

\item{model}{The name of the model for which the main effects should be
returned, one of `dif` or `main`.}

\item{signed}{A logical to use the unsigned DIF effects (`FALSE`) instead of
the signed DIF effects (`TRUE`).}

\item{digits}{A number for rounding.}
}
\value{
The calculated DIF statistics or a vector of item names.
}
\description{
Get values from DIF analyses
}
