% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_extractor.r
\name{GetDifFit}
\alias{GetDifFit}
\alias{GetDIFFit}
\title{Get model fit for DIF analyses}
\usage{
GetDifFit(obj, type, dif = TRUE)

GetDIFFit(obj, type, dif = TRUE)
}
\arguments{
\item{obj}{A data frame with sheets from "dif_dich.xlsx" or "dif_poly.xlsx"
created by [scaling::dif_analysis()].}

\item{type}{The model fit statistic corresponding to the column name in `obj`}

\item{dif}{A logical to return the fit of the DIF model (`TRUE`) instead of
the main effect model (`FALSE`).}
}
\value{
The model fit statistic.
}
\description{
Get model fit for DIF analyses
}
