% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetDifFit}
\alias{GetDifFit}
\alias{GetDIFFit}
\title{Get model fit for DIF analyses}
\usage{
GetDifFit(obj, difvar, type, model = "dif")

GetDIFFit(obj, difvar, type, model = "dif")
}
\arguments{
\item{obj}{A data frame with sheets from "dif_dich_TR.xlsx" or
"dif_poly_TR.xlsx" created by [dif_analysis()].}

\item{difvar}{The DIF variable.}

\item{type}{The model fit statistic corresponding to the column name in `obj`}

\item{model}{The name of the model for which the main effects should be
returned, one of `dif` or `main`.}
}
\value{
The model fit statistic.
}
\description{
Get model fit for DIF analyses
}
\examples{
data("ex1")

# DIF analyses
tmpdir <- tempdir()
dif_analysis(
 resp = ex1$resp,
 vars = ex1$vars,
 select = "dich",
 valid = "valid",
 dif_vars = c("sex", "mig"),
 include_mv = 100,
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 verbose = FALSE,
 warn = FALSE
)

# Import results of DIF analyses
dif <- Import(tmpdir, "dif_dich_TR.xlsx")

# AIC for the variable sex in the DIF model
GetDifFit(dif, "sex", "AIC")

# BIC for the variable mig in the main effect model
GetDifFit(dif, "mig", "BIC", model = "main")

# Clean up generated files
file.remove(paste0(tmpdir, "/dif_dich_sex.xlsx"))
file.remove(paste0(tmpdir, "/dif_dich_mig.xlsx"))
file.remove(paste0(tmpdir, "/dif_dich_TR.xlsx"))
file.remove(paste0(tmpdir, "/dif_dich_models.rds"))
file.remove(paste0(tmpdir, "/dif_dich_summaries.rds"))
}
