% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetDim}
\alias{GetDim}
\title{Summarize the latent correlations in multi-dimensional models}
\usage{
GetDim(obj, model = "dim", stat = median, var = FALSE, digits = 2)
}
\arguments{
\item{obj}{A data frame with sheets from "dimensionality.xlsx"
created by [dim_analysis()].}

\item{model}{A model name which is typically one of uni or dim.}

\item{stat}{A function used to summarize the correlations.}

\item{var}{A logical to summarize the variances (`TRUE`) instead of the
correlations (`FALSE`).}

\item{digits}{A number for rounding.}
}
\value{
A numeric or character vector
}
\description{
Summarize the latent correlations in multi-dimensional models
}
\examples{
data("ex2")

# Dimensionality analyses
tmpdir <- tempdir()
ex2$vars$content <- as.numeric(ex2$vars$content)
dim_analysis(
 resp = ex2$resp,
 vars = ex2$vars,
 select = "mixed",
 valid = "valid",
 dim = "content",
 snodes = 5,# only for speed; model will not converge!
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 verbose = FALSE,
)

# Import results of dimensionality analyses
dim <- Import(tmpdir, "/dimensionality.xlsx")

# Median correlation between factors for variable content
#   Note: not meaningful here because the model did not converge
GetDim(dim, "content", median)

# Largest variance of the dimensions for variable content
#   Note: not meaningful here because the model did not converge
GetDim(dim, "content", max, var = TRUE)

# Clean up generated files
file.remove(paste0(tmpdir, "/dimensionality.xlsx"))
file.remove(paste0(tmpdir, "/dimensionality.rds"))
}
