% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetDim}
\alias{GetDim}
\title{Summarize the latent correlations in multi-dimensional models}
\usage{
GetDim(obj, model = "dim", stat = median, var = FALSE, digits = 2)
}
\arguments{
\item{obj}{A data frame with sheets from "dimensionality.xlsx"
created by [scaling::dim_analysis()].}

\item{model}{A model name which is typically one of uni or dim.}

\item{stat}{A function used to summarize the correlations.}

\item{var}{A logical to summarize the variances (`TRUE`) instead of the
correlations (`FALSE`).}

\item{digits}{A number for rounding.}
}
\value{
A numeric or character vector
}
\description{
Summarize the latent correlations in multi-dimensional models
}
