% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetDimFit}
\alias{GetDimFit}
\title{Get model fit for dimensionality analyses}
\usage{
GetDimFit(obj, model, type)
}
\arguments{
\item{obj}{A data frame with sheets from "dimensionality.xlsx"
created by [scaling::dim_analysis()].}

\item{model}{A column name identifying the model in `obj` which is typically
one of uni or dim.}

\item{type}{A row name identifying the model fit statistic in `obj` which
is typically one of Npars, loglik, AIC, or BIC.}
}
\value{
The model fit statistic.
}
\description{
Get model fit for dimensionality analyses
}
