% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetDist}
\alias{GetDist}
\title{Summarize distractor analyses}
\usage{
GetDist(obj, stat = median, correct = FALSE, item = FALSE, digits = 2)
}
\arguments{
\item{obj}{A data frame with sheets from "distractors_summary.xlsx"
created by [scaling::dis_analysis()].}

\item{stat}{A function used to summarize the distractor correlations.}

\item{correct}{A logical indicating whether to return the results for
distractors (`FALSE`) or correct response options (`TRUE`).}

\item{item}{A logical indicating whether to return the results from `stat`
(`FALSE`) or the item name corresponding to the value of `stat` (`TRUE`).}

\item{digits}{A number for rounding.}
}
\value{
The calculated statistic or a vector of item names.
}
\description{
Summarize distractor analyses
}
