% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetFit}
\alias{GetFit}
\title{Get model fit}
\usage{
GetFit(obj, type, GPCM = FALSE, digits = 2)
}
\arguments{
\item{obj}{A data frame with sheets from "irt_dich.xlsx" or "irt_poly.xlsx"
created by [scaling::irt_analysis()].}

\item{type}{A column name in `obj` identifying the model fit statistic.}

\item{GPCM}{A logical to return the GPCM fit (`TRUE`) instead of the PCM fit
(`FALSE`).}

\item{digits}{A number for rounding.}
}
\value{
The model fit statistic.
}
\description{
Get model fit
}
