% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetMvi}
\alias{GetMvi}
\alias{GetMVI}
\title{Calculate statistics for missing values by item}
\usage{
GetMvi(obj, type, stat, digits = 0)

GetMVI(obj, type, stat, digits = 0)
}
\arguments{
\item{obj}{A list with data frames with sheets from "mv_item.xlsx"
or a single data frame with the sheet "summary" from "mv_item.xlsx" created
by [scaling::mv_item()].}

\item{type}{The type of missing values identifying the row in `obj` which is
typcially one of OM, NR, NV, ND, or ALL.}

\item{stat}{The calculated statistic identifying the column in `obj` which is
typically one of Mean, SD, Median, Min, or Max.}

\item{digits}{A number for rounding.}
}
\value{
A number with the calculated statistic.
}
\description{
Calculate statistics for missing values by item
}
