% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetMvp}
\alias{GetMvp}
\alias{GetMVP}
\title{Calculate percentage of missing values by person}
\usage{
GetMvp(obj, type, value, digits = 0)

GetMVP(obj, type, value, digits = 0)
}
\arguments{
\item{obj}{A list with data frames with sheets from "mv_person.xlsx"
created by \code{\link[=mv_person]{mv_person()}}.}

\item{type}{The type of missing value which is typically one of
OM, NR, NV, ND, or ALL.}

\item{value}{An operator (\code{=}, \code{<}, \code{>}) with a number of missing values.
For example:
\itemize{
\item \code{"=3"} returns the percentage for 3 missing values
\item \code{"<3"} returns the percentage for less than 3 missing values
\item \code{">3"} returns the percentage for more than 3 missing values
}}

\item{digits}{A number for rounding.}
}
\value{
The calculated percentage.
}
\description{
Calculate percentage of missing values by person
}
\examples{
data("ex1")

# Missing analyses by person
tmpdir <- tempdir()
mv_person(
 resp = ex1$resp,
 vars = ex1$vars,
 select = "dich",
 valid = "valid",
 plots = FALSE,
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 warn = FALSE,
 verbose = FALSE,
)

# Import results of missing analyses
mvp <- Import(tmpdir, "/mv_person.xlsx")

# Percentage of respondents with 0 invalid responses
GetMvp(mvp, "NV", "=0")

# Percentage of respondents with more than 5 missing responses with one decimal
GetMvp(mvp, "ALL", ">5", digits = 1)

# Clean up generated files
file.remove(paste0(tmpdir, "/mv_person.xlsx"))
file.remove(paste0(tmpdir, "/mv_person.rds"))
}
