% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetMvp}
\alias{GetMvp}
\alias{GetMVP}
\title{Calculate percentage of missing values by person}
\usage{
GetMvp(obj, type, value, digits = 0)

GetMVP(obj, type, value, digits = 0)
}
\arguments{
\item{obj}{A list with data frames with sheets from "mv_person.xlsx"
created by [scaling::mv_person()].}

\item{type}{The type of missing value which is typically one of
OM, NR, NV, ND, or ALL.}

\item{value}{An operator (=, <, >) with a number of missing values; e.g.,
* =3 returns percentage for 3 missing values
* <3 returns percentage for less than 3 missing values
* >3 returns percentage for more than 3 missing values}

\item{digits}{A number for rounding.}
}
\value{
The calculated percentage.
}
\description{
Calculate percentage of missing values by person
}
