% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetPars}
\alias{GetPars}
\title{Summarize item parameters}
\usage{
GetPars(
  obj,
  type,
  stat = median,
  item = FALSE,
  excl = NULL,
  digits = 2,
  rename_collapsed = TRUE
)
}
\arguments{
\item{obj}{A data frame with sheets from "irt_dich.xlsx" or "irt_poly.xlsx"
created by [irt_analysis()].}

\item{type}{A column name in `obj` identifying an item statistic.}

\item{stat}{A function used to summarize `type` or a string specifying
an operator (=, <, >) with a number; e.g.,
* =3 returns number of type equal to 3
* <3 returns number of type less than 3
* >3 returns number of type greater than 3
 Multiple conditions can be combined using & and |.}

\item{item}{A logical indicating whether to return the results from stat
(`FALSE`) or the item name corresponding to the value of `stat` (`TRUE`).}

\item{excl}{A string specifying an operator (=, <, >) with a number
indicating values to exclude; e.g.,
* =3 excludes number of type equal to 3
* <3 excludes number of type less than 3
* >3 excludes number of type greater than 3
Multiple conditions can be combined using & and |.}

\item{digits}{A number for rounding.}

\item{rename_collapsed}{A boolean to remove the "_collapsed" suffix from
item names.}
}
\value{
A number with the calculated statistic or character vector with
item names.
}
\description{
Summarize item parameters
}
\examples{
\dontrun{
data("ex1")

# IRT analyses
tmpdir <- tempdir()
irt_analysis(
 resp = ex1$resp,
 vars = ex1$vars,
 select = "dich",
 valid = "valid",
 plots = FALSE,
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 warn = FALSE,
 verbose = FALSE,
)

# Import results of IRT analyses
pars <- Import(tmpdir, "/irt_dich.xlsx")

# Smallest percentage of correct responses
GetPars(pars, "correct", min)

# Most difficult item
GetPars(pars, "xsi", max, item = TRUE)

# Items with a difficulty smaller than -2
GetPars(pars, "xsi", "<-2", item = TRUE)

# Largest WMNSQ between 1 and 1.2
GetPars(pars, "WMNSQ", max, excl = "<1|>1.2")

# Clean up generated files
file.remove(paste0(tmpdir, "/irt_dich.xlsx"))
file.remove(paste0(tmpdir, "/irt_dich.rds"))
}
}
