% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetPars}
\alias{GetPars}
\title{Summarize item parameters}
\usage{
GetPars(obj, type, stat = median, item = FALSE, excl = NULL, digits = 2)
}
\arguments{
\item{obj}{A data frame with sheets from "irt_dich.xlsx" or "irt_poly.xlsx"
created by [scaling::irt_analysis()].}

\item{type}{A column name in `obj` identifying an item statistic.}

\item{stat}{A function used to summarize `type` or a string specifying
an operator (=, <, >) with a number; e.g.,
* =3 returns number of type equal to 3
* <3 returns number of type less than 3
* >3 returns number of type greater than 3
 Multiple conditions can be combined using & and |.}

\item{item}{A logical indicating whether to return the results from stat
(`FALSE`) or the item name corresponding to the value of `stat` (`TRUE`).}

\item{excl}{A string specifying an operator (=, <, >) with a number
indicating values to exclude; e.g.,
* =3 excludes number of type equal to 3
* <3 excludes number of type less than 3
* >3 excludes number of type greater than 3
Multiple conditions can be combined using & and |.}

\item{digits}{A number for rounding.}
}
\value{
A number with the calculated statistic or character vector with
item names.
}
\description{
Summarize item parameters
}
