% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_get.r
\name{GetRel}
\alias{GetRel}
\title{Get reliability}
\usage{
GetRel(obj, WLE = FALSE, digits = 2)
}
\arguments{
\item{obj}{A [TAM__taml.mml()] object with a fitted IRT model from
"irt_dich.rds" or "irt_poly.rds" generated by [irt_analysis()].}

\item{WLE}{A logical to return the WLE reliability (`TRUE`) instead of the
EAP reliability (`FALSE`).}

\item{digits}{A number for rounding.}
}
\value{
The reliability.
}
\description{
Get reliability
}
\examples{
data("ex1")

# IRT analyses
tmpdir <- tempdir()
irt_analysis(
 resp = ex1$resp,
 vars = ex1$vars,
 select = "dich",
 valid = "valid",
 plots = FALSE,
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 warn = FALSE,
 verbose = FALSE,
)

# Import results of IRT analyses
irt <- readRDS(paste0(tmpdir, "/irt_dich.rds"))

# EAP reliability
GetRel(irt)

# WLE reliability
GetRel(irt, TRUE)

# Clean up generated files
file.remove(paste0(tmpdir, "/irt_dich.xlsx"))
file.remove(paste0(tmpdir, "/irt_dich.rds"))
}
