% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_extractor.r
\name{GetRel}
\alias{GetRel}
\title{Get reliability}
\usage{
GetRel(obj, WLE = FALSE, digits = 2)
}
\arguments{
\item{obj}{A [TAM__taml.mml()] object with a fitted IRT model from
"irt_dich.rds" or "irt_poly.rds" generated by [scaling::irt_analysis()].}

\item{WLE}{A logical to return the WLE reliability (`TRUE`) instead of the
EAP reliability (`FALSE`).}

\item{digits}{A number for rounding.}
}
\value{
The reliability.
}
\description{
Get reliability
}
