% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_extractor.r
\name{GetVar}
\alias{GetVar}
\title{Get variance of latent proficiency distribution}
\usage{
GetVar(obj, digits = 2)
}
\arguments{
\item{obj}{A [TAM::tam.mml()] object with a fitted IRT model in
"irt_dich.rds" or "irt_poly.rds" generated by [scaling::irt_analysis()].}

\item{digits}{A number for rounding.}
}
\value{
The population variance.
}
\description{
Get variance of latent proficiency distribution
}
