% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{icc_plots}
\alias{icc_plots}
\title{ICC plots}
\usage{
icc_plots(model, path = here::here("Plots"), name_group = NULL)
}
\arguments{
\item{model}{list; return object of irt_model()}

\item{path}{string; defines path to folder where plots shall be saved}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}
}
\description{
Create ICC plots for IRT models.
}
