% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{irt_all}
\alias{irt_all}
\title{IRT analyses - all in one function}
\usage{
irt_all(
  resp,
  vars,
  items,
  irt_type,
  position = NULL,
  Q = NULL,
  icc_plots = TRUE,
  wright_map = TRUE,
  path_plots = here::here("Plots"),
  name_table = NULL,
  path_table = here::here("Tables"),
  digits = 2,
  return_results = TRUE,
  name_data = NULL,
  path_data = here::here("Data")
)
}
\arguments{
\item{resp}{data.frame. contains:
(1) the responses. y in {0, 1} for binary data and y in
  {0, 1, ... k-1} for polytomous responses with k categories
(2) ID_t: column indicating ID of participants}

\item{vars}{data.frame. contains all competence items as rows,
and at least the following variables:
  character vector named "items"; contains the names of the items.
  boolean vector; indicates which items to use for analysis.}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{irt_type}{character string; either "dich" for dichotomous analysis
or "poly" for polytomous analysis.}

\item{position}{character string. contains name of position variable.}

\item{Q}{matrix with one column and ncol(resp) rows. Binary items are
scored with 1, polytomous items are scored with 0.5.
Divergent scoring decisions can be incorporated by this matrix
(e.g., item 7 is scored with 0.75 by assigning 0.75 to the seventh row).}

\item{icc_plots}{boolean; indicates whether to create and save ICC plots.}

\item{wright_map}{boolean; indicates whether to create and save Wright Map.}

\item{path_plots}{character. contains name of path for plots.}

\item{name_table}{string with name of table file that shall be saved (including file type);
if left empty, table will not be saved..}

\item{path_table}{character. contains name of path for table.}

\item{digits}{number of decimals for rounding}

\item{return_results}{boolean. indicates whether to return results.}

\item{name_data}{string with name of data file that shall be saved (including file type);
if left empty, data will not be saved..}

\item{path_data}{character. contains name of path for data.}
}
\description{
Perform 1PL and 2PL analyses for binary or PCM and GPCM analyses for
polytomous data, generate summary and plots.
}
