% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{irt_analysis}
\alias{irt_analysis}
\title{IRT analyses - all in one function}
\usage{
irt_analysis(
  resp,
  vars,
  items,
  position,
  valid = NULL,
  irt_type,
  scoring = NULL,
  icc_plots = FALSE,
  wright_map = FALSE,
  path_plots = here::here("Plots"),
  name_table = NULL,
  name_steps = NULL,
  path_tables = here::here("Tables"),
  name_data = NULL,
  path_data = here::here("Data"),
  overwrite_table = FALSE,
  digits = 2,
  return_results = TRUE
)
}
\arguments{
\item{resp}{data.frame. contains:
(1) the responses. y in {0, 1} for binary data and y in
  {0, 1, ... k-1} for polytomous responses with k categories
(2) ID_t: column indicating ID of participants}

\item{vars}{data.frame. contains all competence items as rows,
and at least the following variables:
  character vector named "items"; contains the names of the items.
  boolean vector; indicates which items to use for analysis.}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{position}{character string. contains name of position variable.}

\item{valid}{character string. defines name of boolean variable in resp,
indicating (in)valid cases.}

\item{irt_type}{character string; either "dich" for dichotomous analysis
or "poly" for polytomous analysis.}

\item{scoring}{numeric vector; scoring factor to be applied to loading matrix;
can be NULL for Rasch model; if NULL for PCM model,
scoring matrix is reconstructed from item names.}

\item{icc_plots}{boolean; indicates whether to create and save ICC plots.}

\item{wright_map}{boolean; indicates whether to create and save Wright Map.}

\item{path_plots}{character. contains name of path for plots.}

\item{name_table}{string with name of summary table file that shall be saved (including file type);
if left empty, table will not be saved.}

\item{name_steps}{string with name of steps table file that shall be saved (including file type);
if left empty, table will not be saved; only available for partial credit models.}

\item{path_tables}{character. contains name of path for tables.}

\item{name_data}{string with name of data file that shall be saved (including file type);
if left empty, data will not be saved.}

\item{path_data}{character. contains name of path for data.}

\item{overwrite_table}{boolean; indicates whether to overwrite existing file when saving table.}

\item{digits}{number of decimals for rounding}

\item{return_results}{boolean. indicates whether to return results.}
}
\description{
Perform 1PL and 2PL analyses for binary or PCM and GPCM analyses for
polytomous data, generate summary and plots.
}
