% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{irt_analysis}
\alias{irt_analysis}
\title{IRT analyses - all in one function}
\usage{
irt_analysis(
  resp,
  vars,
  select,
  valid = NULL,
  mvs = NULL,
  irt_type,
  scoring = NULL,
  plots = FALSE,
  save = TRUE,
  print = TRUE,
  return = FALSE,
  path_plots = here::here("Plots"),
  path_table = here::here("Tables"),
  path_results = here::here("Results"),
  overwrite = FALSE,
  digits = 2,
  name_group = NULL,
  verbose = FALSE,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{irt_type}{string; either "dich" (dichotomous analysis) or "poly"
(polytomous analysis)}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; can be NULL for
Rasch model}

\item{plots}{logical; whether plots shall be created and saved to hard drive}

\item{save}{logical; whether results shall be saved to hard drive}

\item{print}{logical; whether results shall be printed to console}

\item{return}{logical; whether results shall be returned}

\item{path_plots}{string; defines path to folder where plots shall be saved}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{path_results}{string; defines path to folder where results shall be saved}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{digits}{integer; number of decimals for rounding}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}

\item{verbose}{logical; whether to print processing information to console}

\item{warn}{logical; whether to print warnings (should be set to TRUE to
avoid problems with the data structure)}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\value{
list with all results
}
\description{
Perform 1PL and 2PL analyses for binary or PCM and GPCM analyses for
polytomous data, generate summary and plots.
}
