% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{irt_model}
\alias{irt_model}
\title{IRT analyses}
\usage{
irt_model(
  resp,
  vars,
  items,
  valid = NULL,
  irtmodel,
  scoring = NULL,
  path = here::here("Data"),
  filename = NULL,
  verbose = FALSE,
  return_results = TRUE
)
}
\arguments{
\item{resp}{data.frame. contains:
(1) the responses. y in {0, 1} for binary data and y in
  {0, 1, ... k-1} for polytomous responses with k categories
(2) ID_t: column indicating ID of participants}

\item{vars}{data.frame. contains all competence items as rows,
and at least the following variables:
  character vector named "items"; contains the names of the items.
  boolean vector; indicates which items to use for analysis.}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{valid}{character string. defines name of boolean variable in resp,
indicating (in)valid cases.}

\item{irtmodel}{character. "1PL" for Rasch, "2PL" for 2PL, "PCM2" for PCM and
"GPCM" for GPCM analyses.}

\item{scoring}{numeric vector; scoring factor to be applied to loading matrix;
can be NULL for Rasch model; if NULL for PCM model,
scoring matrix is reconstructed from item names.}

\item{path}{folder path for data}

\item{filename}{string with name of file that shall be saved (including file type).}

\item{verbose}{logical. If verbose == TRUE information about the estimation
progress is printed to the console}

\item{return_results}{boolean. indicates whether to return results.}
}
\value{
(if return_results = TRUE) a list of:
  mod: tam.mml; estimated item response model
  fit: data.frame; item fit (WMNSQ) of the model
  pars: data.frame; item difficulties and scoring parameters with respective standard errors
  mfit: data.frame; (adjusted) Yen's Q3 statistics of the model
  wle: data.frame; WLEs for the model
  wle_rel: numeric; WLE reliability of the model
  info_crit: data.frame with information criteria of the model
}
\description{
Perform Rasch or 2PL analyses for binary or PCM or GPCM analyses for
polytomous data.
}
