% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{irt_model}
\alias{irt_model}
\title{IRT analyses}
\usage{
irt_model(
  resp,
  vars,
  select,
  valid = NULL,
  mvs = NULL,
  irtmodel,
  scoring = NULL,
  verbose = FALSE,
  path = here::here("Results"),
  filename = NULL,
  xsi.fixed = NULL,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{irtmodel}{string; "1PL" for Rasch, "2PL" for 2PL, "PCM2" for PCM and
"GPCM" for GPCM analysis}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; can be NULL for
Rasch model}

\item{verbose}{logical; whether to print processing information to console}

\item{path}{string; defines path to folder where results shall be saved}

\item{filename}{string; defines name of file that shall be saved}

\item{xsi.fixed}{matrix; contains fixed (linked) item parameters; optional}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\value{
(if return_results = TRUE) a list of:
  mod: tam.mml; estimated item response model
  fit: data.frame; item fit (WMNSQ) of the model
  pars: data.frame; item difficulties and scoring parameters with respective standard errors
  mfit: data.frame; (adjusted) Yen's Q3 statistics of the model
  wle: data.frame; WLEs for the model
  wle_rel: numeric; WLE reliability of the model
  info_crit: data.frame with information criteria of the model
}
\description{
Perform Rasch or 2PL analyses for binary or PCM or GPCM analyses for
polytomous data.
}
