% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{irt_model_fit}
\alias{irt_model_fit}
\title{IRT model fit}
\usage{
irt_model_fit(
  model_1p,
  model_2p,
  overwrite = FALSE,
  save = TRUE,
  path = here::here("Tables"),
  name_group = NULL
)
}
\arguments{
\item{model_1p}{list; return object of irt_model(); 1PL / PCM analysis}

\item{model_2p}{list; return object of irt_model(); 2PL / GPCM analysis}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{save}{logical; whether results shall be saved to hard drive}

\item{path}{string; defines path to folder where table shall be saved}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}
}
\value{
data.frame with N, Deviance, AIC, BIC and number of parameters for
both models.
}
\description{
Create table with model fit for 1 parameter and 2 parameter IRT model.
}
