% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{irt_summary}
\alias{irt_summary}
\title{Summary of IRT analysis}
\usage{
irt_summary(
  resp,
  vars,
  results,
  position,
  disc = NULL,
  path = here::here("Tables"),
  filename = NULL,
  valid = NULL,
  digits = 2,
  overwrite = FALSE,
  return_table = TRUE
)
}
\arguments{
\item{resp}{data.frame. contains:
(1) the responses. y in {0, 1} for binary data and y in
{0, 1, ... k-1} for polytomous responses with k categories
(2) ID_t: column indicating ID of participants}

\item{vars}{data.frame. contains information about all competence items
and includes the following columns:
  items: character indicating names of items.
  final: logical indicating whether item is a final item
  position: contains position per item. if different for
  different groups, several variables necessary.}

\item{results}{list. contains results from IRT analysis with one parameter
(e.g., Rasch analysis)}

\item{position}{character string. contains name of position variable}

\item{disc}{list. contains results from IRT analysis with two parameters
(e.g., 2PL analysis), to include item discrimination}

\item{path}{character. defines name of path for table.}

\item{filename}{character. defines name for excel document. if NULL (default),
the table will not be saved.}

\item{valid}{character string. defines name of boolean variable in resp,
indicating (in)valid cases.}

\item{digits}{integer; how many digits after rounding}

\item{overwrite}{boolean; indicates whether to overwrite existing file when saving table.}

\item{return_table}{logical; whether resulting table should be returned}
}
\value{
a data.frame containing the item name, position, N, percentage correct,
  item difficulty, SE, WMNSQ, t, rit, item discrimination, Q3.
}
\description{
Create table with results of IRT analysis.
}
