% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{link_item_parameters}
\alias{link_item_parameters}
\title{Link item parameters with known link constant}
\usage{
link_item_parameters(
  xsi,
  const,
  vars,
  select,
  scoring = NULL,
  return_steps = FALSE,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{xsi}{data.frame; xsi parameters as returned by TAM::tam.mml()}

\item{const}{numeric; link constant}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that
indicates which items to use for the analysis}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to the link constant (e.g., only
half the link constant is needed if a PC item is scored 0.5)}

\item{return_steps}{logical; should steps be returned (not re-estimated in
subsequent analysis with TAM::tam.mml()). Defaults to FALSE.}

\item{warn}{logical; whether to print warnings}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\value{
matrix in xsi.fixed format needed for TAM::tam.mml()
}
\description{
Link item parameters with known link constant
}
