% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{link_samples}
\alias{link_samples}
\title{Link two samples using mean-mean linking}
\usage{
link_samples(
  resp_previous,
  resp_current,
  resp_link_sample = NULL,
  vars,
  select_current,
  select_previous,
  select_common,
  valid,
  scoring,
  anchors = NULL,
  longitudinal = TRUE,
  pid = "ID_t",
  wid = "wle",
  mvs = NULL
)
}
\arguments{
\item{resp_previous}{data.frame with responses of first measurement wave,
a person identifier and a variable indicating valid cases; may also
include a WLE as given in wid}

\item{resp_current}{data.frame with responses of second measurement wave,
a person identifier and a variable indicating valid cases}

\item{resp_link_sample}{data.frame with responses of link sample,
a person identifier and a variable indicating valid cases; variables of
the link sample have to be adapted to the item names and the item scoring
in the previous and current measurement}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select_current}{character; contains name of logical variable in vars
identifying the item set for the current measurement time point}

\item{select_previous}{character; contains name of logical variable in vars
identifying the item set for the previous measurement time point}

\item{select_common}{character; contains name of logical variable in vars
identifying the item set for the link sample}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to the link constant (e.g., only
half the link constant is needed if a PC item is scored 0.5)}

\item{anchors}{character vector with link items
if NULL, all common items are used}

\item{longitudinal}{do within cohort linking (TRUE) or between cohort
linking (FALSE)}

\item{pid}{variable name used as person identifier}

\item{wid}{variable name used as WLE identifier in first measurement wave}

\item{mvs}{named integer vector; contains user-defined missing values}
}
\value{
list: wle = linked WLEs for second measurement with
                       standard errors
                 const   = linking constant
                 const.err = linking error
                 anchors = used anchor items
                 k = number of anchor items
                 xsi = linked item parameters
}
\description{
Link two samples using mean-mean linking
}
