% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{link_wles}
\alias{link_wles}
\title{Link WLEs using a known linking constant while correcting for selection}
\usage{
link_wles(
  wle_previous,
  wle_current,
  const,
  pid = "ID_t",
  wid = "wle",
  ids = NULL,
  use_longitudinal_subsample = TRUE
)
}
\arguments{
\item{wle_previous}{data.frame with WLEs at first measurement wave}

\item{wle_current}{data.frame with WLEs at second measurement wave}

\item{const}{linking constant}

\item{pid}{character; variable name used as person identifier}

\item{wid}{character; variable name used as WLE identifier; optional
as vector containing different WLE identifier names}

\item{ids}{vector of person identifiers for longitudinal sample;
if null, all cases with the same pid in wle_previous and wle_current
will be used}

\item{use_longitudinal_subsample}{boolean indicating whether linking should
be based on the entire samples given in wle_previous and wle_current or
only on the longitudinal sample with common pids; defaults to TRUE}
}
\value{
linked WLEs at second measurement wave
}
\description{
Link WLEs using a known linking constant while correcting for selection
}
