% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{linking}
\alias{linking}
\title{Mean/mean linking of current and previous study}
\usage{
linking(
  resp_previous,
  resp_current,
  resp_link_sample = NULL,
  vars,
  valid = NULL,
  select_previous,
  select_current,
  select_common,
  scoring = "scoring",
  overwrite = FALSE,
  mvs = NULL,
  maxiter,
  snodes,
  verbose,
  anchors = NULL,
  longitudinal = TRUE,
  path_table = here::here("Tables"),
  path_results = here::here("Results"),
  pid = "ID_t",
  wid = "wle",
  dim = "link",
  print = TRUE,
  save = TRUE,
  return = FALSE,
  diff_threshold = 0.5,
  anchor_item_names = NULL
)
}
\arguments{
\item{resp_previous}{data.frame with responses of first measurement wave,
a person identifier and a variable indicating valid cases; may also
include a WLE as given in wid}

\item{resp_current}{data.frame with responses of second measurement wave,
a person identifier and a variable indicating valid cases}

\item{resp_link_sample}{data.frame with responses of link sample,
a person identifier and a variable indicating valid cases; variables of
the link sample have to be adapted to the item names and the item scoring
in the previous and current measurement}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments: contains an
item identifier for all items in the link sample and a dimension identifier
for the time points; please note that, in the anchor items design, the
common items between the time points must be named the same and, therefore,
not be doubled in vars}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{select_previous}{character; contains name of logical variable in vars
identifying the item set for the previous measurement time point}

\item{select_current}{character; contains name of logical variable in vars
identifying the item set for the current measurement time point}

\item{select_common}{character; indicates the variable in vars
selecting the items that were administered in the link sample (anchor
groups design) or at both time points (anchor items design -- note that
anchor items only appear once!)}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to the link constant (e.g., only
half the link constant is needed if a PC item is scored 0.5)}

\item{overwrite}{logical; whether to overwrite existing file when saving
table}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{maxiter}{max iterations as passed to the TAM function}

\item{snodes}{snodes as passed to the TAM function}

\item{verbose}{verbose as passed to the TAM function}

\item{anchors}{character vector with link items; if NULL, all common items
are used in the anchor items design and adjusted according to the link
analyses}

\item{longitudinal}{do within cohort linking (TRUE) or between cohort
linking (FALSE)}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{path_results}{string; defines path to folder where results shall be
saved}

\item{pid}{variable name used as person identifier}

\item{wid}{variable name used as WLE identifier in first measurement wave}

\item{dim}{character; indicates the variable in vars defining the dimension
of the items; anchor groups design: the time points form distinct
dimensions; anchor items design: the common items across time points and
the time point unique items form unique dimensions, respectively}

\item{print}{logical; whether results shall be printed to console}

\item{save}{logical; whether results shall be saved to hard drive}

\item{return}{logical; whether results shall be returned}

\item{diff_threshold}{numeric; threshold under which DIF in common link items
is accepted; defaults to .5}

\item{anchor_item_names}{character matrix; contains the old (column 1) and
new (column 2) item names of the anchor items for the TR table in the
anchor item design; NOT to be confused with the argument anchors that
is a vector of the anchor items CURRENT name for the analysis; if anchors
is not NULL, anchor_item_names may also only contain the same item name
mapping as anchors}
}
\description{
Mean/mean linking of current and previous study
}
