% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{min_val}
\alias{min_val}
\title{Select sample with a minimum number of valid values}
\usage{
min_val(resp, vars, select, min.val = NULL, invalid = NULL)
}
\arguments{
\item{resp}{data.frame with item responses}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{min.val}{minimum number of valid values; if negative, set to the default of 3}

\item{invalid}{vector of invalid values (if not specified,
function defaults to NA and negative values)}
}
\value{
logical vector with length = nrow(resp), indicating whether case is valid
}
\description{
Select sample with a minimum number of valid values
}
