% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_val}
\alias{min_val}
\title{Select sample with a minimum number of valid values}
\usage{
min_val(x, vars = NULL, min.val = 3, invalid = NA, append = FALSE, warn = TRUE)
}
\arguments{
\item{x}{data.frame with responses}

\item{vars}{variables to check for valid values;}

\item{min.val}{minimum number of valid values;
if negative, set to the default of 3}

\item{invalid}{vector of invalid values}

\item{append}{boolean indicating whether to filter data.frame (FALSE)
or append new variable (TRUE)}

\item{warn}{print warnings}
}
\value{
filtered data.frame or list of results
}
\description{
Select sample with a minimum number of valid values
}
