% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_val}
\alias{min_val}
\title{Select sample with a minimum number of valid values}
\usage{
min_val(resp, vars, items, min.val = NULL, invalid = NA, warn = TRUE)
}
\arguments{
\item{resp}{data.frame with responses}

\item{vars}{variables to check for valid values;}

\item{min.val}{minimum number of valid values;
if negative, set to the default of 3}

\item{invalid}{vector of invalid values}

\item{warn}{print warnings}
}
\value{
boolean vector with length = nrow(resp),
indicating whether case is valid.
}
\description{
Select sample with a minimum number of valid values
}
