% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_item.R
\name{mv_item}
\alias{mv_item}
\title{Missing values analysis by item - all in one function}
\usage{
mv_item(
  resp,
  vars,
  select,
  valid = NULL,
  position = NULL,
  grouping = NULL,
  stages = NULL,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, MD = -54, AZ = -21),
  labels_mvs = c(ALL = "total missing items", OM = "omitted items", NV =
    "not valid items", NR = "not reached items", TA =
    "missing items due to test abortion", UM = "unspecific missing items", ND =
    "not determinable items", MD = "items missing by design", AZ =
    "missing items due to 'Angabe zurueckgesetzt'"),
  missing_by_design = -54,
  plots = FALSE,
  print = TRUE,
  save = TRUE,
  return = FALSE,
  path_results = "Results",
  path_table = "Tables",
  path_plots = "Plots/Missings_by_item",
  suf_item_names = FALSE,
  show_all = TRUE,
  name_grouping = "test version",
  overwrite = FALSE,
  name_group = NULL,
  color = NULL,
  digits = 3,
  warn = TRUE,
  verbose = TRUE,
  labels_legend = NULL
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; all responses ∈ ℕ0; user-defined missing values;
includes all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{position}{(named) character vector; defines name(s) of integer
variable(s) in vars that indicate position of items; if groups with differing
item positions in testlets exist, then vector must be named with names of
groups (as in "grouping") as names of elements and names of variables as elements}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{stages}{character vector; contains names of stage variables in resp and vars,
only applicable in multistage tests (otherwise NULL)}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{labels_mvs}{named character vector; contains labels for user-defined
missing values to use them in plot titles and printed results}

\item{missing_by_design}{integer; user defined missing value for missing by
design}

\item{plots}{logical; whether plots shall be created and saved to hard drive}

\item{print}{logical; whether results shall be printed to console}

\item{save}{logical; whether results shall be saved to hard drive}

\item{return}{logical; whether results shall be returned}

\item{path_results}{string; defines path to folder where results shall be saved}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{path_plots}{string; defines path to folder where plots shall be saved}

\item{suf_item_names}{logical; whether to output SUF item names in the .xlsx file
for items with collapsed categories}

\item{show_all}{logical; whether whole sample shall be included as a "group"
(only applicable when grouping exists)}

\item{name_grouping}{string; name of the grouping variable (e.g. test version)
for title and legend of plots (only needed when grouping exists)}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{name_group}{string; defines name of group used in analysis (e.g. 'settingA')}

\item{color}{character scalar or vector; defines color(s) of the bar in plots}

\item{digits}{integer; number of decimals for rounding}

\item{warn}{logical; whether to print warnings (should be better set to true)}

\item{verbose}{logical; whether to print processing information to console}

\item{labels_legend}{character vector; contains legend labels}
}
\value{
(if return = TRUE) list with results:
           list: percentages for each item
           summary: summary statistics across items
           summary_table: table with summary statistics for TR.
}
\description{
Missing values analysis by item - all in one function
}
