% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{mv_person}
\alias{mv_person}
\title{Missing values by person analysis}
\usage{
mv_person(
  resp,
  vars,
  items = "final",
  grouping = NULL,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  name_data = NULL,
  path_data = here::here("Data"),
  return_table = TRUE,
  name_table = NULL,
  path_table = here::here("Tables"),
  plots = FALSE,
  name_plots = "Missing_responses_by_person",
  path_plots = here::here("Plots/Missing_Responses/by_person"),
  color_plots = NULL,
  show.all_plots = FALSE,
  lbls_plots = c(ALL = "total missing", OM = "omitted", NV = "not valid", NR =
    "not reached", TA = "test aborted", UM = "unspecific missing", ND =
    "not determinable", NAd = "not administered", AZ = "Angabe zurueckgesetzt"),
  min.val = 3,
  digits = 2,
  warn = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{resp}{data.frame with item responses (user-defined missing values)
and logical grouping variables, named after the grouping
factor (e.g. "easy" and "difficult" in the case of testlet
difficulties) --> must be in separate columns!}

\item{vars}{data.frame. contains all competence items as rows,
with the column "items" indicating the names of the items.}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to items in analysis.}

\item{grouping}{character vector. contains names of groups (e.g. 'easy and 'difficult')
as items in data.frame resp.}

\item{mvs}{named vector with definition of user-defined missing values}

\item{name_data}{string with name of data file that shall be saved (including type of file);
if NULL, no file will be saved.}

\item{path_data}{folder path for file if it shall be saved}

\item{return_table}{boolean. indicates whether table shall be returned.}

\item{name_table}{string with name of table file that shall be saved (excluding type of file);
if NULL, no file will be saved.}

\item{path_table}{folder path for file if it shall be saved}

\item{plots}{boolean. indicates whether plotsshall be created.}

\item{name_plots}{string with base of name for plot files (excluding type of file)}

\item{path_plots}{folder path for plots}

\item{color_plots}{bar color}

\item{show.all_plots}{logical; only needed when groups exist, indicates whether
plots shall also include the whole sample as a "group"}

\item{lbls_plots}{lables for different types of user-defined missing values}

\item{min.val}{minimum number of valid values;
if negative, set to the default of 3}

\item{digits}{number of decimals for rounding}

\item{warn}{boolean. indicates whether to print a warning if NAs were found in resp.}

\item{verbose}{print progress}
}
\value{
if return_table is set to TRUE, table with missing values per person and
for each missing value type (in percentage)
}
\description{
Missing values by person analysis
}
