% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{mv_person}
\alias{mv_person}
\title{Missing values by person analysis}
\usage{
mv_person(
  resp,
  vars,
  select,
  valid = NULL,
  grouping = NULL,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, MD = -54, AZ = -21),
  labels_mvs = c(ALL = "total missing items", OM = "omitted items", NV =
    "not valid items", NR = "not reached items", TA =
    "missing items due to test abortion", UM = "unspecific missing items", ND =
    "not determinable items", MD = "items missing by design", AZ =
    "missing items due to 'Angabe zurueckgesetzt'"),
  missing_by_design = -54,
  color = NULL,
  plots = FALSE,
  print = TRUE,
  save = TRUE,
  return = FALSE,
  path_results = here::here("Results"),
  path_table = here::here("Tables"),
  path_plots = here::here("Plots/Missing_Responses/by_person"),
  show_all = TRUE,
  overwrite = FALSE,
  name_group = NULL,
  name_grouping = "test version",
  labels_legend = NULL,
  digits = 3,
  warn = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; all responses ∈ ℕ0; user-defined missing values;
includes all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{labels_mvs}{named character vector; contains labels for user-defined
missing values to use them in plot titles and printed results}

\item{missing_by_design}{integer; user defined missing value for missing by
design}

\item{color}{character scalar or vector; defines color(s) of the bar in plots}

\item{plots}{logical; whether plots shall be created and saved to hard drive}

\item{print}{logical; whether results shall be printed to console}

\item{save}{logical; whether results shall be saved to hard drive}

\item{return}{logical; whether results shall be returned}

\item{path_results}{string; defines path to folder where results shall be saved}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{path_plots}{string; defines path to folder where plots shall be saved}

\item{show_all}{logical; whether whole sample shall be included as a "group"
(only applicable when grouping exists)}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{name_group}{string; defines name of group used in analysis (e.g. 'settingA')}

\item{name_grouping}{string; name of the grouping variable (e.g. test version)
for title and legend of plots (only needed when grouping exists)}

\item{labels_legend}{character vector; contains legend labels}

\item{digits}{integer; number of decimals for rounding}

\item{warn}{logical; whether to print warnings (should be set to TRUE to
avoid problems with the data structure)}

\item{verbose}{logical; whether to print processing information to console}
}
\value{
(if return = TRUE) list with results:
            mv_p: list with missing values per person and each missing value type (in percentage)
            summary: table with summary statistics for TR.
}
\description{
Missing values by person analysis
}
