% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_item.R
\name{mvi_analysis}
\alias{mvi_analysis}
\title{Percentage of missing responses by item}
\usage{
mvi_analysis(
  resp,
  vars,
  items,
  position = NULL,
  grouping = NULL,
  show.all = FALSE,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  path = here::here("Data"),
  filename = NULL,
  min.val = 3,
  digits = 2,
  warn = TRUE,
  return_results = TRUE
)
}
\arguments{
\item{resp}{data.frame. contains item responses (integer with user-defined missing values)
and grouping variables (boolean).}

\item{vars}{data.frame. contains all competence items as rows,
and at least the following variables:
  character vector named "items"; contains the names of the items.
  boolean vector; indicates which items to use for analysis.
  integer vector; contains position for each item; if grouping
    exists and positions differ between groups, several variables necessary.}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{position}{(named) character vector. contains name(s) of variable(s) in
vars that indicate the position of subitems;
if grouping with differing positions in testlets,
then for each group one variable name is necessary,
as well as names for the vector that represent names of groups
and must be identical with items in variable "grouping"
(e.g. position = c(easy = "position_easy", diff = "position_diff"))}

\item{grouping}{character vector. contains names of groups (e.g. 'easy and 'difficult')
as used in resp.}

\item{show.all}{boolean; only needed when groups exist, indicates whether
plots shall also include the whole sample as a "group"}

\item{mvs}{named vector with definition of user-defined missing values.}

\item{path}{folder path for data}

\item{filename}{string with name of file that shall be saved}

\item{min.val}{minimum number of valid values;
if negative, set to the default of 3}

\item{digits}{number of decimals for rounding}

\item{warn}{boolean whether to print a warning if NAs were found in resp}

\item{return_results}{boolean. indicates whether to return results.}
}
\value{
list with percentages:
           mvlist: percentages for each item
           mvsum: summary statistics across items
}
\description{
Percentage of missing responses by item
}
