% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_item.R
\name{mvi_plots}
\alias{mvi_plots}
\title{Plot percentages of user-defined missing values by item position}
\usage{
mvi_plots(
  vars,
  items,
  mv_i = NULL,
  resp = NULL,
  position = NULL,
  grouping = NULL,
  show.all = FALSE,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  path = here::here("Plots/Missing_Responses/by_item"),
  filename = "Missing_responses_by_item",
  color = NULL,
  verbose = TRUE,
  min.val = 3,
  digits = 2,
  warn = TRUE
)
}
\arguments{
\item{vars}{data.frame. contains all competence items as rows,
and at least the following variables:
character vector named "items"; contains the names of the items.
boolean vector; indicates which items to use for analysis.}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}

\item{mv_i}{list; return object of mvi_analysis(). mvi_analysis$list must contain
for each group and all groups together the name of the
user-defined missing values that are to be
analyzed and the names of the items as its row names.}

\item{resp}{data.frame. contains item responses (integer with user-defined missing values)
and grouping variables (boolean).}

\item{position}{character vector. contains names of position variables
(one element per grouping, same order necessary)}

\item{grouping}{character vector. contains names of groups (e.g. 'easy and 'difficult')
as used in resp.}

\item{show.all}{boolean; only needed when groups exist, indicates whether
plots shall also include the whole sample as a "group"}

\item{mvs}{named vector; contains definition of user-defined missing
values that shall be analyzed in the plot}

\item{path}{folder path for plots}

\item{filename}{string with name of file that shall be saved (excluding type of file)}

\item{color}{character vector. defines bar color(s); if the data contains subgroups,
one color per subgroup must be specified}

\item{verbose}{logical; print information to console}

\item{min.val}{minimum number of valid values;
if negative, set to the default of 3}

\item{digits}{number of decimals for rounding}

\item{warn}{boolean whether to print a warning if NAs were found in resp}
}
\description{
Plots the percentages of not reached missing values for each item by its
item position. If not otherwise specified, the order in which the items are
given to the function is taken as the order the items appeared in the test.
}
