% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_item.R
\name{mvi_plots}
\alias{mvi_plots}
\title{Plot percentages of user-defined missing values by item position}
\usage{
mvi_plots(
  resp,
  vars,
  select,
  valid,
  position,
  stages = NULL,
  grouping = NULL,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, MD = -54, AZ = -21),
  labels_mvs = c(ALL = "total missing items", OM = "omitted items", NV =
    "not valid items", NR = "not reached items", TA =
    "missing items due to test abortion", UM = "unspecific missing items", ND =
    "not determinable items", MD = "items missing by design", AZ =
    "missing items due to 'Angabe zurueckgesetzt'"),
  missing_by_design = -54,
  path = "Plots/Missings_by_item",
  name_group = NULL,
  name_grouping = "test version",
  show_all = TRUE,
  labels_legend = NULL,
  color = NULL,
  digits = 3,
  verbose = TRUE,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; all responses ∈ ℕ0; user-defined missing values;
includes all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{position}{(named) character vector; defines name(s) of integer
variable(s) in vars that indicate position of items; if groups with differing
item positions in testlets exist, then vector must be named with names of
groups (as in "grouping") as names of elements and names of variables as elements}

\item{stages}{character vector; contains names of stage variables in resp and vars,
only applicable in multistage tests (otherwise NULL)}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{labels_mvs}{named character vector; contains labels for user-defined
missing values to use them in plot titles and printed results}

\item{missing_by_design}{integer; user defined missing value for missing by
design}

\item{path}{string; defines path to folder where plots shall be saved}

\item{name_group}{string; defines name of group used in analysis (e.g. 'settingA')}

\item{name_grouping}{string; name of the grouping variable (e.g. 'test version')
for title and legend of plots (only needed when grouping exists)}

\item{show_all}{logical; whether whole sample shall be included as a "group"
(only applicable when grouping exists)}

\item{labels_legend}{character vector; contains legend labels}

\item{color}{character scalar or vector; defines color(s) of the bar in plots}

\item{digits}{integer; number of decimals for rounding}

\item{verbose}{logical; whether to print processing information to console}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\description{
Plots the percentages of not reached missing values for each item by its
item position. If not otherwise specified, the order in which the items are
given to the function is taken as the order the items appeared in the test.
}
