% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_item.R
\name{mvi_plots}
\alias{mvi_plots}
\title{Plot percentages of user-defined missing values by item position}
\usage{
mvi_plots(
  mv_i,
  vars,
  select,
  grouping = NULL,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, MD = -54, AZ = -21),
  labels_mvs = c(ALL = "total missing items", OM = "omitted items", NV =
    "not valid items", NR = "not reached items", TA =
    "missing items due to test abortion", UM = "unspecific missing items", ND =
    "not determinable items", MD = "items missing by design", AZ =
    "missing items due to 'Angabe zurueckgesetzt'"),
  missing_by_design = -54,
  path = here::here("Plots/Missing_Responses/by_item"),
  name_group = NULL,
  name_grouping = "test version",
  show_all = TRUE,
  labels_legend = NULL,
  color = NULL,
  verbose = TRUE,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{mv_i}{list; return object of mvi_analysis()}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{labels_mvs}{named character vector; contains labels for user-defined
missing values to use them in plot titles and printed results}

\item{missing_by_design}{integer; user defined missing value for missing by
design}

\item{path}{string; defines path to folder where plots shall be saved}

\item{name_group}{string; defines name of group used in analysis (e.g. 'settingA')}

\item{name_grouping}{string; name of the grouping variable (e.g. 'test version')
for title and legend of plots (only needed when grouping exists)}

\item{show_all}{logical; whether whole sample shall be included as a "group"
(only applicable when grouping exists)}

\item{labels_legend}{character vector; contains legend labels}

\item{color}{character scalar or vector; defines color(s) of the bar in plots}

\item{verbose}{logical; whether to print processing information to console}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\description{
Plots the percentages of not reached missing values for each item by its
item position. If not otherwise specified, the order in which the items are
given to the function is taken as the order the items appeared in the test.
}
