% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_item.R
\name{mvi_table}
\alias{mvi_table}
\title{Write table of user defined missing values per item (and position)}
\usage{
mvi_table(
  vars,
  items,
  mv_i = NULL,
  resp = NULL,
  valid = NULL,
  position = NULL,
  grouping = NULL,
  show_all = FALSE,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  filename = NULL,
  path = here::here("Tables"),
  digits = 2,
  warn = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{items}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{mv_i}{list; return object of mvi_analysis()}

\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; all responses ∈ ℕ0; user-defined missing values;
includes all variables that are further defined in the function arguments}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{position}{(named) character vector; defines name(s) of integer
variable(s) in vars that indicate position of items; if groups with differing
item positions in testlets exist, then vector must be named with names of
groups (as in "grouping") as names of elements and names of variables as elements}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{show_all}{logical; whether whole sample shall be included as a "group"
(only applicable when grouping exists)}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{filename}{string; defines name of table that shall be saved}

\item{path}{string; defines path to folder where table shall be saved}

\item{digits}{integer; number of decimals for rounding}

\item{warn}{logical; whether to print a warning if NAs were found in resp}

\item{overwrite}{logical; whether to overwrite existing file when saving table}
}
\value{
table with results
}
\description{
Plots the percentages of not reached missing values for each item by its
item position. If not otherwise specified, the order in which the items are
given to the function is taken as the order the items appeared in the test.
}
