% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_item.R
\name{mvi_table}
\alias{mvi_table}
\title{Write table of user defined missing values per item (and position)}
\usage{
mvi_table(
  mv_i,
  vars,
  select,
  grouping = NULL,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  save = TRUE,
  path = here::here("Tables"),
  overwrite = FALSE,
  test = TRUE,
  warn = TRUE
)
}
\arguments{
\item{mv_i}{list; return object of mvi_analysis()}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{save}{logical; whether results shall be saved to hard drive}

\item{path}{string; defines path to folder where table shall be saved}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{test}{logical; whether to test data structure (should be set to TRUE)}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}
}
\value{
table with results
}
\description{
Plots the percentages of not reached missing values for each item by its
item position. If not otherwise specified, the order in which the items are
given to the function is taken as the order the items appeared in the test.
}
