% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{mvp_analysis}
\alias{mvp_analysis}
\title{Percentage of missing responses by person}
\usage{
mvp_analysis(
  resp,
  vars,
  items = "final",
  grouping = NULL,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  filename = NULL,
  path = here::here("Data"),
  min.val = 3,
  digits = 2,
  warn = TRUE,
  return_results = TRUE
)
}
\arguments{
\item{resp}{data.frame with item responses (user-defined missing values)
and logical grouping variables, named after the grouping
factor (e.g. "easy" and "difficult" in the case of testlet
difficulties) --> must be in separate columns!}

\item{vars}{data.frame. contains all competence items as rows,
with the column "items" indicating the names of the items.}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to items in analysis.}

\item{grouping}{character vector. contains names of groups (e.g. 'easy and 'difficult')
as itemsd in data.frame resp.}

\item{mvs}{named vector with definition of user-defined missing values}

\item{filename}{string with name of file that shall be saved (including type of file);
if NULL, no file will be saved.}

\item{path}{folder path for file if it shall be saved}

\item{min.val}{minimum number of valid values;
if negative, set to the default of 3}

\item{digits}{number of decimals for rounding}

\item{warn}{boolean. indicates whether to print a warning if NAs were found in resp}

\item{return_results}{boolean. indicates whether to return results.}
}
\value{
if return is set to TRUE, list with missing values per person and
for each missing value type (in percentage)
}
\description{
Percentage of missing responses by person
}
