% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{mvp_analysis}
\alias{mvp_analysis}
\title{Percentage of missing responses by person}
\usage{
mvp_analysis(
  resp,
  vars,
  select,
  valid = NULL,
  grouping = NULL,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, MD = -54, AZ = -21),
  missing_by_design = -54,
  save = TRUE,
  path = here::here("Results"),
  name_group = NULL,
  digits = 3,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; all responses ∈ ℕ0; user-defined missing values;
includes all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{grouping}{character vector; contains for each group a name of a logical
variable in resp and vars that indicates to which group belongs a person or
an item}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{missing_by_design}{integer; user defined missing value for missing by
design}

\item{save}{logical; whether results shall be saved to hard drive}

\item{path}{string; defines path to folder where results shall be saved}

\item{name_group}{string; defines name of group used in analysis (e.g. 'settingA')}

\item{digits}{integer; number of decimals for rounding}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\value{
list with missing values per person and for each missing value type
(in percentage).
}
\description{
Percentage of missing responses by person
}
