% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_person.R
\name{mvp_table}
\alias{mvp_table}
\title{Write table of user defined missing values per item (and position)}
\usage{
mvp_table(
  mv_p = NULL,
  grouping = NULL,
  resp = NULL,
  vars = NULL,
  valid = NULL,
  mvs = c(OM = -97, NV = -95, NR = -94, TA = -91, UM = -90, ND = -55, NAd = -54, AZ =
    -21),
  filename = NULL,
  path = here::here("Tables"),
  overwrite = FALSE,
  return_table = FALSE
)
}
\arguments{
\item{mv_p}{list; return object of mv_person().}

\item{grouping}{character vector. contains names of grouping variables.}

\item{resp}{data.frame. contains item responses (user-defined mvs)}

\item{vars}{data.frame, contains information about variables
(for minimum columns see function "mv_person")}

\item{valid}{character string. defines name of boolean variable in resp,
indicating (in)valid cases.}

\item{mvs}{named vector; contains definition of user-defined missing
values that shall be analyzed in the plot}

\item{filename}{string with name of file that shall be saved (excluding type of file);
if NULL, no file will be saved.}

\item{path}{folder path for plots}

\item{overwrite}{boolean; indicates whether to overwrite existing file when saving table.}

\item{return_table}{boolean. indicates whether to return table.}
}
\value{
if return is set to TRUE, table with missing values per person and
for each missing value type (in percentage)
}
\description{
Plots the percentages of not reached missing values for each item by its
item position. If not otherwise specified, the order in which the items are
given to the function is taken as the order the items appeared in the test.
}
