% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{n_valid}
\alias{n_valid}
\title{Number of valid cases}
\usage{
n_valid(dat, valid = NULL)
}
\arguments{
\item{dat}{data.frame; contains sociodemographic and booklet variables}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}
}
\value{
character string with answer.
}
\description{
Number of valid cases
}
