% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEPSroutines.R
\docType{package}
\name{NEPSroutines-package}
\alias{NEPSroutines}
\alias{NEPSroutines-package}
\title{NEPSroutines}
\description{
This package contains a collection of functions for the
scaling of NEPS competence data. It provides standardized scaling
routines for one- and two-parametric logistic models with binary and
polytomous responses. It supports analyses of descriptive information,
missing values, distractors, dimensionality, and differential item
functioning. Also, it can generate item and respondent scores
for the scientific use files in the NEPS. Additionally, several
functions support general data preparation and consistency checks of the
data. Finally, the package includes a Quarto extension and functionalities
for the creation of semi-automated technical reports.
}
\details{
Scaling NEPS data using R in general and NEPSroutines in particular is
documented using \strong{vignettes} within this package.
Start with the \strong{introductory vignette} by using the function \code{open_guide()}.
}
\author{
\strong{Maintainer}: Jana Welling \email{jana.welling@lifbi.de}

Authors:
\itemize{
  \item Jana Welling \email{jana.welling@lifbi.de}
  \item Anna Scharl
  \item Tanja Kutscher \email{tanja.kutscher@lifbi.de}
  \item Timo Gnambs \email{timo.gnambs@lifbi.de}
}

Other contributors:
\itemize{
  \item Christian Diegel [contributor]
}

}
\keyword{internal}
