% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{only_valid}
\alias{only_valid}
\title{Select only valid cases}
\usage{
only_valid(resp, valid = NULL)
}
\arguments{
\item{resp}{data.frame with responses}

\item{valid}{character string. defines name of boolean variable in resp,
indicating (in)valid cases.}
}
\value{
data.frame as resp, but only with valid cases
}
\description{
Select only valid cases
}
