% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{pc_scoring}
\alias{pc_scoring}
\title{Score partial credit items}
\usage{
pc_scoring(resp, subtasks, mvs = NULL)
}
\arguments{
\item{resp}{data.frame; contains original item responses}

\item{subtasks}{character vector; contains variable names of subtasks}

\item{mvs}{integer vector; contains user-defined missing values}
}
\description{
Score partial credit items
}
