% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{pc_scoring}
\alias{pc_scoring}
\title{Score partial credit items}
\usage{
pc_scoring(resp, poly_items, mvs = NULL, warn = TRUE)
}
\arguments{
\item{resp}{data.frame; contains original item responses}

\item{poly_items}{list; contains character vector with subitems for each
polytomous item, name of the vector is the name of the polytomous item (e.g.
poly_items = list(poly1 = c("subitem1", "subitem2"), poly2 = c("subitem1", "subitem2")))}

\item{mvs}{integer vector; contains user-defined missing values}

\item{warn}{logical; print warnings}
}
\value{
resp including unscored (raw) and scored items
}
\description{
Score partial credit items
}
