% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pos_new}
\alias{pos_new}
\title{Calculate new position variable with only a set of variables}
\usage{
pos_new(vars, items = "final", position = NULL)
}
\arguments{
\item{vars}{data.frame with information about all items}

\item{items}{character. contains name of variable in vars (boolean) that
indicates which variables to use.}

\item{position}{(named) character vector. contains name(s) of variable(s) in
vars that indicate the position of subitems;
if grouping, then for each group one variable name is necessary;
names represent names of groups (e.g. easy = "position_easy")}
}
\value{
data.frame as input, with one extra variable indicating position of chosen variables.
}
\description{
Calculate new position variable with only a set of variables
}
