% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_resp}
\alias{prepare_resp}
\title{Prepare resp for analysis}
\usage{
prepare_resp(
  resp,
  valid = NULL,
  without_valid = FALSE,
  vars = NULL,
  items = NULL,
  convert = FALSE
)
}
\arguments{
\item{resp}{data.frame with responses}

\item{valid}{character string. defines name of boolean variable in resp,
indicating (in)valid cases.}

\item{without_valid}{boolean; indicates whether not to check valid cases}

\item{vars}{data.frame. contains information about all competence items
and includes the following columns:
  items: character indicating names of items.
  variable indicated by argument 'items'}

\item{items}{character. contains name of variable (boolean) in vars that
indicates which items to use for analysis.}
}
\value{
data.frame as resp, but only with valid cases
}
\description{
Prepare resp for analysis
}
