% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{print_dif_summary}
\alias{print_dif_summary}
\title{Print DIF summaries to console}
\usage{
print_dif_summary(diflist, res)
}
\arguments{
\item{diflist}{list; return object of dif_model(); with main and dif model}

\item{res}{list; return object of dif_summaries()}
}
\description{
Print DIF summaries to console
}
