% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{props_by_version}
\alias{props_by_version}
\title{Table with item properties by test version}
\usage{
props_by_version(
  vars,
  select,
  grouping,
  properties,
  labels = NULL,
  save = FALSE,
  overwrite = FALSE,
  path = here::here("Tables"),
  warn = TRUE
)
}
\arguments{
\item{vars}{data.frame; contains information about the competence items}

\item{save}{logical; whether to save the table in Excel}

\item{overwrite}{logical; whether to overwrite an existing table with the same name}

\item{path}{string; defines path for saving the table}

\item{versions}{string; defines name of variable in vars that identifies test versions}

\item{props}{string; defines name of variable in vars that identifies item properties}

\item{versions_lbls}{named character vector; links a label each value of props
(e.g. labels[[props]] = c(prop1 = 1, prop2 = 2))}
}
\value{
table with item properties by test version
}
\description{
Table with item properties by test version
}
