% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_suf_and_sum_scores.R
\name{set_labels}
\alias{set_labels}
\title{Set labels of items in SUF}
\usage{
set_labels(suf, vars, items_suf, competence, wle_name, linked)
}
\arguments{
\item{suf}{data.frame; SUF to be (re)labelled}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{items_suf}{character; indicates the logical variable in vars which
contains the item names designated for the SUF (original scored items
before the first IRT analyses)}

\item{competence}{character; name of tested competence for labelling the
data (capitalized!)}

\item{wle_name}{character; name of the wle -- WITHOUT extension (e.g.,
reg4 instead of reg4_sc1 or mag12 instead of mag12_sc1u)}

\item{linked}{logical; whether the test is linked to a previous assessment}
}
\value{
suf with labels
}
\description{
Set labels of items in SUF
}
