% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{steps_analysis}
\alias{steps_analysis}
\title{Step analysis}
\usage{
steps_analysis(
  pcm_model,
  digits = 3,
  save = TRUE,
  overwrite = FALSE,
  path = here::here("Tables"),
  name_group = NULL
)
}
\arguments{
\item{pcm_model}{list; return object of irt_model(); PCM analysis}

\item{digits}{integer; number of decimals for rounding}

\item{save}{logical; whether results shall be saved to hard drive}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{path}{string; defines path to folder where table shall be saved}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}
}
\value{
a data.frame containing the step parameters and SEs for each step.
}
\description{
Create table with pcm_model of step analysis.
}
