% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_suf_and_sum_scores.R
\name{sum_scores}
\alias{sum_scores}
\title{create sum scores}
\usage{
sum_scores(resp, vars, items)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in {0, 1} for binary data and y in {0, 1, ... k-1} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person
identifier and all variables that are further defined in the function
arguments; if special person sets are of interest, these persons have to
be selected in resp beforehand}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'items' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{items}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}
}
\value{
integer vector of length nrow(resp) containing the row sums for the
  specified item set
}
\description{
Sum scores for PC items are only counted as 1 if all subitems are solved
  correctly
}
