% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{summarize_dif_analysis}
\alias{summarize_dif_analysis}
\title{summarize DIF analysis}
\usage{
summarize_dif_analysis(
  dif_models,
  dif_vars,
  dif_threshold = 0.5,
  print = TRUE,
  save = TRUE,
  overwrite = FALSE,
  path_results = here::here("Results"),
  path_table = here::here("Tables")
)
}
\arguments{
\item{dif_models}{return object of conduct_dif_analysis()}

\item{dif_vars}{character vector; contains the variable names to be tested
for DIF (e.g., "gender")}

\item{dif_threshold}{numeric scalar; indicates absolute threshold of
problematic DIF (defaults to 0.5)}

\item{print}{logical; whether results shall be printed to console}

\item{save}{logical; whether results shall be saved to hard drive}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{path_results}{string; defines path to folder where results shall be saved}

\item{path_table}{string; defines path to folder where tables shall be saved}
}
\value{
a list of dif summaries for each input entry in dif_models
}
\description{
summarize DIF analysis
}
