% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{summarize_link_dif}
\alias{summarize_link_dif}
\title{Summarize link DIF results}
\usage{
summarize_link_dif(
  dif_summary = NULL,
  mod_current = NULL,
  mod_previous = NULL,
  items_previous,
  items_current,
  anchors = NULL
)
}
\arguments{
\item{dif_summary}{return object of dif_model() (anchor groups design)}

\item{mod_current}{return object of irt_model() for current assessment
(anchor items design)}

\item{mod_previous}{return object of irt_model() for previous assessment
(anchor items design)}

\item{items_previous}{character vector; item names of previous assessment}

\item{items_current}{character vector; item names of current assessment}

\item{anchors}{character vector with link items; if NULL, all common items
are used}
}
\description{
Summarize link DIF results
}
