% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{summarize_link_dif}
\alias{summarize_link_dif}
\title{Summarize link DIF results}
\usage{
summarize_link_dif(
  mod_curr,
  mod_prev,
  mod_link = NULL,
  items_curr,
  items_prev,
  items_link = NULL,
  anchors = NULL,
  dif_threshold = 0.5
)
}
\arguments{
\item{mod_curr}{return object of irt_model() for current assessment}

\item{mod_prev}{return object of irt_model() for previous assessment}

\item{mod_link}{return object of irt_model() for link study}

\item{items_curr}{character vector; item names of current assessment}

\item{items_prev}{character vector; item names of previous assessment}

\item{items_link}{character vector; item names of link study}

\item{anchors}{character; data.frame with two columns including the link
items; for anchor item designs the first column refers to the previous
measurement time point and the second column to the current measurement
time point; for anchor group designs the first column refers to the main
sample (for both time points) and the second column refers to the link
sample; if NULL, all common items are used}

\item{dif_threshold}{numeric; threshold under which DIF in common link items
is accepted; defaults to 0.5}
}
\value{
list with DIF results
}
\description{
Summarize link DIF results
}
