% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblCode}
\alias{TblCode}
\title{Create a table with the analysis code}
\usage{
TblCode(vars, select, collapsed = NULL, tbl = TRUE, special = FALSE)
}
\arguments{
\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{character; indicates the logical variable in vars which
contains the item names of the polytomous items}

\item{collapsed}{A data frame with the collapsed item categories generated
by [collapse_response_categories()].}

\item{tbl}{A logical to return the table (`TRUE`) instead of the code (`FALSE`).}

\item{special}{A logical for special schools (`TRUE`) or regular schools
(`FALSE`).}
}
\value{
A `knitr::knit_child()` object or a vector.
}
\description{
Create a table with the analysis code
}
\examples{
\dontrun{
data(ex2)

TblCode(ex2$vars, "mixed", tbl = FALSE)
}
}
