% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblDif}
\alias{TblDif}
\alias{TblDIF}
\title{Create table for DIF analyses}
\usage{
TblDif(
  obj,
  footnote = NULL,
  excl = NULL,
  colnames1 = NULL,
  colnames2 = NULL,
  width = 1.4,
  size = 10
)

TblDIF(
  obj,
  footnote = NULL,
  excl = NULL,
  colnames1 = NULL,
  colnames2 = NULL,
  width = 1.4,
  size = 10
)
}
\arguments{
\item{obj}{A list with data frames with sheets from "dif_poly_TR.xlsx"
or a data frame with sheet "estimates" from "dif_poly_TR.xlsx" created by
[scaling::dif_analysis()].}

\item{footnote}{A table note.}

\item{excl}{A vector of column names to exclude.}

\item{colnames1}{A vector of names for the DIF variables included in the
first line of the table heading.}

\item{colnames2}{A vector of groups for the DIF variables included in the
second line for the table heading.}

\item{width}{The  widths of the columns; if a single values is given, it
corresponds to the first column; otherwise the number of values must
correspond to the number of columns in `obj`.}

\item{size}{The general font size.}
}
\value{
A flextable.
}
\description{
Create table for DIF analyses
}
