% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblDifFit}
\alias{TblDifFit}
\alias{TblDIFFit}
\title{Create table with fit statistics for DIF analyses}
\usage{
TblDifFit(
  obj,
  footnote = NULL,
  excl = NULL,
  label = NULL,
  size = 12,
  width = 0.9
)

TblDIFFit(
  obj,
  footnote = NULL,
  excl = NULL,
  label = NULL,
  size = 12,
  width = 0.9
)
}
\arguments{
\item{obj}{A list with data frames with sheets from "dif_poly_TR.xlsx"
or a data frame with sheet "estimates" from "dif_poly_TR.xlsx" created by
[scaling::dif_analysis()].}

\item{footnote}{A table note.}

\item{excl}{A vector of column names to exclude.}

\item{label}{A vector of names for the DIF variables.}

\item{size}{The general font size.}

\item{width}{The  widths of the columns; if a single values is given, it
corresponds to the first column; otherwise the number of values must
correspond to the number of columns in `obj`.}
}
\value{
A flextable.
}
\description{
Create table with fit statistics for DIF analyses
}
