% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblDim}
\alias{TblDim}
\title{Create correlation table for multi-dimensional models}
\usage{
TblDim(
  obj,
  model,
  rownames = NULL,
  colnames = NULL,
  footnote = NULL,
  size = 12,
  width = 3
)
}
\arguments{
\item{obj}{A data frame with sheets from "dimensionality.xlsx"
created by [dim_analysis()].}

\item{model}{The model name, typically `uni` for the unidimensional
reference model and the name of the facet variable.}

\item{rownames}{A vector of labels for the rows.}

\item{colnames}{A vector of labels for the columns.}

\item{footnote}{A table note.}

\item{size}{The general font size.}

\item{width}{The column widths; if a single value is given, it refers to the
first column; otherwise the number of values must correspond to the number of
columns in `obj`.}
}
\value{
A flextable.
}
\description{
Create correlation table for multi-dimensional models
}
\examples{
\dontrun{
data(ex2)

# Dimensionality analyses
tmpdir <- tempdir()
ex2$vars$content <- as.numeric(ex2$vars$content)
dim_analysis(
 resp = ex2$resp,
 vars = ex2$vars,
 select = "mixed",
 valid = "valid",
 dim = "content",
 snodes = 5,# only for speed; model will not converge!
 print = FALSE,
 save = TRUE,
 return = FALSE,
 path_results = tmpdir,
 path_table = tmpdir,
 overwrite = TRUE,
 verbose = FALSE,
)

# Import results of dimensionality analyses
dim <- Import(tmpdir, "/dimensionality.xlsx")

# Default correlation table
TblDim(dim, "content")

# Correlation table with labels
TblDim(dim, "content",
       rownames = c("Units", "Change", "Space", "Data"))
}
}
