% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblItemFacets}
\alias{TblItemFacets}
\title{Create a table with facets for each item}
\usage{
TblItemFacets(
  vars,
  select,
  facets,
  position = NULL,
  footnote = NULL,
  size = 12,
  width = 1.9,
  lbl = NULL,
  rename_collapsed = TRUE
)
}
\arguments{
\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{A named vector of variable names in `vars` identifying
different groups of items for which the item property should be counted. The
name of the vector gives the column headings of the table.}

\item{facets}{Variable names in `vars` identifying the item facets.}

\item{position}{The variable name in `vars` giving the item position.}

\item{footnote}{A table note.}

\item{size}{The general font size.}

\item{width}{The widths of the columns; if a single value is provided, it
will be applied to all columns; otherwise the length must correspond to the
number of columns in `obj`.}

\item{lbl}{Long names of item facets to be displayed in the table.}

\item{rename_collapsed}{A boolean to remove the "_collapsed" suffix from
item names.}
}
\value{
A flextable.
}
\description{
Create a table with facets for each item
}
\examples{
\dontrun{
data(ex2)

# Table with content areas
TblItemFacets(
  ex2$vars,
  select = "mixed",
  facets = c("Content area" = "content")
)

# Table with content areas, response formats, and item position
TblItemFacets(
  ex2$vars,
  select = "mixed",
  facets = c("Content area" = "content", "Response format" = "type"),
  position = "pos",
  footnote = "MC = simple multiple-choice, CMC = complex multiple-choice"
)
}
}
