% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblItemProps}
\alias{TblItemProps}
\title{Create a table with item properties}
\usage{
TblItemProps(
  vars,
  select,
  prop,
  propname = " ",
  footnote = NULL,
  na.rm = TRUE,
  size = 12,
  width = NULL,
  formats = NULL,
  ...
)
}
\arguments{
\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{A named vector of variable names in \code{vars} identifying
different groups of items for which the item property should be counted. The
name of the vector gives the column headings of the table.}

\item{prop}{A vector with the variable name in \code{vars} identifying the item
property such as response format, text function, or cognitive function.}

\item{propname}{A name for the item property to be used in the table heading.}

\item{footnote}{A table note.}

\item{na.rm}{A logical to remove empty properties (rows) and empty groups (columns).}

\item{size}{The general font size.}

\item{width}{The widths of the columns; if a single value is provided, it
will be applied to all columns; otherwise the length must correspond to the
number of columns in \code{obj}.}

\item{formats}{Long names of item properties to be displayed in the table.}

\item{...}{Further arguments passed to \code{\link[=Tbl]{Tbl()}}; \code{align} and \code{hline} are set
internally and will error if also passed here.}
}
\value{
A flextable.
}
\description{
Create a table with item properties
}
\examples{
\dontrun{
data(ex2)
data(ex3)

# Table with content areas for a single group
TblItemProps(
  ex2$vars,
  select = c("Number of items" = "mixed"),
  prop = "content",
  propname = "Content area"
)

# Number of response format for three booklets with a footnote
ex3$vars$mixed1 <- ex3$vars$mixed & ex3$vars$booklet1
ex3$vars$mixed2 <- ex3$vars$mixed & ex3$vars$booklet2
ex3$vars$mixed3 <- ex3$vars$mixed & ex3$vars$booklet3
TblItemProps(
  ex3$vars,
  select = c("Booklet 1" = "mixed1", "Booklet 2" = "mixed2",
             "Booklet 3" = "mixed3"),
  prop = "type",
  propname = "Response format",
  footnote = "The study administered three difficulty-tiered booklets."
)
}
}
