% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblItemProps}
\alias{TblItemProps}
\title{Create a table with item properties}
\usage{
TblItemProps(
  vars,
  select,
  prop,
  propname = "",
  footnote = NULL,
  na.rm = TRUE,
  size = 12,
  width = NULL,
  formats = c(MC = "Simple multiple-choice items", CMC = "Complex multiple-choice items",
    SR = "Short constructed responses", MA = "Matching items", TET =
    "Text-enrichment tasks", HL = "Highlighting tasks", finding =
    "Finding information in the text", conclusion = "Drawing text-related conclusions",
    reflecting = "Reflecting and assessing", information = "Information text",
    instruction = "Instruction text", advertising = "Advertising text", commenting =
    "Commenting or argumenting text", literary = "Literary text", 
     change =
    "Change and relationship", data = "Data and chance", units = "Units and measuring",
    space = "Space and shape")
)
}
\arguments{
\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{A named vector of variable names in `vars` identifying
different groups of items for which the item property should be counted. The
name of the vector gives the column headings of the table.}

\item{prop}{A vector with the variable name in `vars` identifying the item
property such as response format, text function, or cognitive function.}

\item{propname}{A name for the item property to be used in the table heading.}

\item{footnote}{A table note.}

\item{na.rm}{A logical to remove empty properties.}

\item{size}{The general font size.}

\item{width}{The widths of the columns; if a single value is provided, it
will be applied to all columns; otherwise the length must correspond to the
number of columns in `obj`.}

\item{formats}{Long names of item properties to be displayed in the table.}
}
\value{
A flextable.
}
\description{
Create a table with item properties
}
