% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblMvi}
\alias{TblMvi}
\alias{TblMVI}
\title{Create a table for missing responses by item}
\usage{
TblMvi(
  obj,
  select = NULL,
  footnote = NULL,
  sort = "position",
  size = 12,
  width = NULL,
  excl = c("N_administered", "ND", "ALL")
)

TblMVI(
  obj,
  select = NULL,
  footnote = NULL,
  sort = "position",
  size = 12,
  width = NULL,
  excl = c("N_administered", "ND", "ALL")
)
}
\arguments{
\item{obj}{A list with data frames with sheets from "mv_item.xlsx"
or a data frame with sheet "summary" from "mv_item.xlsx" created by
[scaling::mv_item()].}

\item{select}{An optional name of a specific group to select.}

\item{footnote}{An optional table note.}

\item{sort}{A column name to indicate by which column to sort.}

\item{size}{The general font size.}

\item{width}{The widths of the columns; if a single value is provided, it
will be applied to all columns; otherwise the length must correspond to the
number of columns in `obj`.}

\item{excl}{A vector of column names to exclude.}
}
\value{
A flextable.
}
\description{
Create a table for missing responses by item
}
