% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblPars}
\alias{TblPars}
\title{Create table with item parameters}
\usage{
TblPars(
  obj,
  footnote = NULL,
  excl = c("N_administered"),
  size = 10,
  width = 0.5
)
}
\arguments{
\item{obj}{A list with data frames with sheets from "irt_poly.xlsx"
or a data frame with sheet "summary" from "irt_poly.xlsx" created by
[scaling::irt_analysis()].}

\item{footnote}{A table note.}

\item{excl}{A vector of column names to exclude.}

\item{size}{The general font size.}

\item{width}{The widths of the columns; if a single value is provided, it
will be applied to all columns; otherwise the length must correspond to the
number of columns in `obj`.}
}
\value{
A flextable.
}
\description{
Create table with item parameters
}
