% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_tables.r
\name{TblSteps}
\alias{TblSteps}
\title{Create table with step parameters}
\usage{
TblSteps(obj, footnote = NULL, size = 10, width = 1)
}
\arguments{
\item{obj}{A list with data frames with sheets from "irt_poly.xlsx"
or sheet "steps" from "irt_poly.xlsx" created by [scaling::irt_analysis()].}

\item{footnote}{A table note.}

\item{size}{The general font size.}

\item{width}{The widths of the columns; if a single value is provided, it
will be applied to all columns; otherwise the length must correspond to the
number of columns in `obj`.}
}
\value{
A flextable.
}
\description{
Create table with step parameters
}
