% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{wright_map}
\alias{wright_map}
\title{Wright maps}
\usage{
wright_map(results, name, path = here::here("Plots"))
}
\arguments{
\item{results}{list. Contains results from IRT analysis as returned by function
'irt_analysis'.}

\item{name}{character. Part of the filename for the directory of the plots,
following "ICCs_for_" (e.g., the kind of IRT analysis: "1PL").}

\item{path}{character. contains name of path for plots.}
}
\description{
Create Wright maps for IRT models.
}
